### 2022-07-07 #4553 @editorialbot
[PRE REVIEW]: oxDNA: coarse-grained simulations of nucleic acids made simple
**Submitting author:** @lorenzo-rovigatti (<a href="http://orcid.org/0000-0001-5017-2829">Lorenzo Rovigatti</a>)
**Repository:** https://github.com/lorenzo-rovigatti/oxDNA
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** v3.2.2
