### 2022-07-07 #4553 @danielskatz
👋 @lorenzo-rovigatti - note that your paper does not compile.

The error includes "found a tab character that violate indentation while scanning a plain scalar at line 10 column 11"

Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
