### 2022-07-07 #4553 @danielskatz
Also in the bib, I see a couple of places the authors include "and others" - can you please add all the authors in the bib file, and let the system figure out how many to list.
