### 2022-07-07 f6007db
Remove the JOSS paper from the repo
