+++ joss/paper.md
@@ -1,67 +1,0 @@
----
-title: 'oxDNA: coarse-grained simulations of nucleic acids made simple'
-tags:
-  - C++
-  - Python
-  - DNA
-  - DNA nanotechnology
-  - Molecular simulations
-authors:
-  - name: Erik Poppleton
-  	orcid:  0000-0002-5146-5970 
-    affiliation: 1
-  - name: Michael Matthies
-  	orcid: 
-    affiliation: 1
-  - name: Flavio Romano
-  	orcid: 
-  	affiliation: 2
-  - name: Petr Šulc
-   	orcid: 0000-0003-1565-6769
-   	affiliation: 1
-  - name: Lorenzo Rovigatti^[Corresponding author]
-    orcid: 0000-0001-5017-2829
-    affiliation: "3, 4"
-affiliations:
- - name: School of Molecular Sciences and Center for Molecular Design and Biomimetics, The Biodesign Institute, Arizona State University, USA
-   index: 1
- - name: Dipartimento di Scienze Molecolari e Nanosistemi, Universitá Ca Foscari di Venezia, Italy
-   index: 2
- - name: Department of Physics, Sapienza University of Rome, Italy
-   index: 3
- - name: CNR-ISC UoS Sapienza, Rome, Italy
-   index: 4
-date: 08 May 2022
-bibliography: paper.bib
----
-
-# Summary
-
-The fields of DNA and RNA nanotechnology have progressed from pioneering, proof-of-principle experiments to fully-fledged applications in material science, biology and medicine. These applications exploit the intrinsic programmability of nucleic acids to generate nano- and even micro-scale structures with tailored properties. However, the design of the DNA/RNA sequences that self-assemble into a desired structure is not straightforward and often relies on expensive trial-and-error experimental protocols. A complementary approach is provided by computer simulations, which can model biomacromolecules at different levels of detail, ranging from atomistic to continuous, and can be leveraged to investigate the whole range of time- and length-scales relevant for applications.
-
-# Statement of need
-
-The simulation of nucleic acids has become an important tool from the fundamental point of view to understand how these biomacromolecules behave and, from an application standpoint, to predict their behaviour under specific conditions[@DNA_simulation_review]. The specific model that one should use, and therefore the level of detail with which DNA and RNA are to be described, depends on the time and length scales of interest: at one end of the spectrum is the quantum chemistry modelling, which can be used to probe the microscopic properties of a small number of nucleotides[@quantum_DNA], while at the other end are continuum descriptions based on polymer theories such as the worm-like chain model[@WLC] that can be used to study the behaviour of long DNA strands. A middle road is provided by coarse-grained models that describe nucleic acids at the nucleotide level[@oxDNA_review]. At this level of detail the oxDNA and oxRNA models has become popular choices to investigate the dynamics, thermodynamics and self-assembly kinetics of DNA and RNA systems[@oxRNA; @oxDNA_primer].
-
-# Functionality
-
[…]
