+++ joss/paper.bib
@@ -1,263 +1,0 @@
-@Article{quantum_DNA,
-author ="Šponer, Jiří and Riley, Kevin E. and Hobza, Pavel",
-title  ="Nature and magnitude of aromatic stacking of nucleic acid bases",
-journal  ="Phys. Chem. Chem. Phys.",
-year  ="2008",
-volume  ="10",
-issue  ="19",
-pages  ="2595-2610",
-publisher  ="The Royal Society of Chemistry",
-doi  ="10.1039/B719370J",
-url  ="http://dx.doi.org/10.1039/B719370J",
-abstract  ="This review summarises recent advances in quantum chemical calculations of base-stacking forces in nucleic acids. We explain in detail the very complex relationship between the gas-phase base-stacking energies{,} as revealed by quantum chemical (QM) calculations{,} and the highly variable roles of these interactions in nucleic acids. This issue is rarely discussed in quantum chemical and physical chemistry literature. We further extensively discuss methods that are available for base-stacking studies{,} complexity of comparison of stacking calculations with gas phase experiments{,} balance of forces in stacked complexes of nucleic acid bases{,} and the relation between QM and force field descriptions. We also review all recent calculations on base-stacking systems{,} including details analysis of the B-DNA stacking. Specific attention is paid to the highest accuracy QM calculations{,} to the decomposition of the interactions{,} and development of dispersion-balanced DFT methods. Future prospects of computational studies of base stacking are discussed."}
-}
-
-@article{WLC,
-  title = {Twist-bend coupling and the statistical mechanics of the twistable wormlike-chain model of DNA: Perturbation theory and beyond},
-  author = {Nomidis, Stefanos K. and Skoruppa, Enrico and Carlon, Enrico and Marko, John F.},
-  journal = {Phys. Rev. E},
-  volume = {99},
-  issue = {3},
-  pages = {032414},
-  numpages = {16},
-  year = {2019},
-  month = {Mar},
-  publisher = {American Physical Society},
-  doi = {10.1103/PhysRevE.99.032414},
-  url = {https://link.aps.org/doi/10.1103/PhysRevE.99.032414}
-}
-
-@article{DNA_simulation_review,
-	title = {Multiscale simulation of DNA},
-	journal = {Current Opinion in Structural Biology},
-	volume = {37},
-	pages = {29-45},
-	year = {2016},
-	note = {Theory and simulation • Macromolcular machines},
-	issn = {0959-440X},
-	doi = {https://doi.org/10.1016/j.sbi.2015.11.011},
-	url = {https://www.sciencedirect.com/science/article/pii/S0959440X15001761},
-	author = {Pablo D Dans and Jürgen Walther and Hansel Gómez and Modesto Orozco},
-	abstract = {DNA is not only among the most important molecules in life, but a meeting point for biology, physics and chemistry, being studied by numerous techniques. Theoretical methods can help in gaining a detailed understanding of DNA structure and function, but their practical use is hampered by the multiscale nature of this molecule. In this regard, the study of DNA covers a broad range of different topics, from sub-Angstrom details of the electronic distributions of nucleobases, to the mechanical properties of millimeter-long chromatin fibers. Some of the biological processes involving DNA occur in femtoseconds, while others require years. In this review, we describe the most recent theoretical methods that have been considered to study DNA, from the electron to the chromosome, enriching our knowledge on this fascinating molecule.}
-}
-
-@ARTICLE{oxDNA_primer,
-	AUTHOR={Sengar, A. and Ouldridge, T. E. and Henrich, O. and Rovigatti, L. and Šulc, P.},	 
[…]
