### 2022-07-07 5daa2e7
Remove the CI action that compiles the JOSS paper
