### 2022-07-07 #4553 @danielskatz
@lorenzo-rovigatti - I see a bunch of references (most of them) that are missing a space before them.

In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
