+++ docs/source/install.md
@@ -47,6 +47,8 @@ The following options pertain to `oxpy`:
 * `-DPython=ON` Enables Python bindings
 * `-DOxpySystemInstall=On` By default `oxpy` is installed in the current user's home directory. By enabling this option `oxpy` will be installed as a system-wide package. It may require superuser privileges.
 
+If you are on your own machine or you installed Python via Anaconda, the `-DOxpySystemInstall=On` option should be set to install `oxpy` and `oxDNA_analysis_tools` on your `PATH`.  If you are on a shared system (for example and HPC cluster which isn't using Conda), perform the `cmake` command without this option.  `make install` will probably (depends on `pip` settings) install the two libraries in `$HOME/.local/bin/`, which is not on the `PATH` by default and you will have to add it (if you have not already for another program).  In our testing, Conda-managed versions of Python correctly installed the libraries in the current environment's `site-packages` folder if `-DOxpySystemInstall=On` was set, even on shared systems.
+
 ### make targets
 
 * `make` compiles oxDNA
@@ -61,10 +63,6 @@ When compiling with the Python bindings enabled CMake will sometimes choose the 
 ```bash
 cmake .. -DPython=ON -DPYTHON_INCLUDE_DIRS=/path/to/python/include/dir -DPYTHON_EXECUTABLE=/path/to/python/binary
 ```
-
-If you get an error regarding the number of bytes in the `numpy.array` header, this happens when the version of Numpy on your system doesn't match the version that pip downloads from PyPi when installing `oat` with its isolated environment (most commonly because you installed Numpy using Conda which tends to be a few versions behind PyPi). To fix this, try to manually install `oat` with:
-  
-   `python -m pip install ./analysis --no-build-isolation`
 	
 ## Testing
 
