### 2022-07-08 98ac417
Make it possible to access the last_conf observable from python
This commit allows to do something like this to print the current
configuration whenever is required:

	obs = manager.config_info().get_observable_by_id("last_conf")
    manager.update_CPU_data_structures()
    conf = obs.get_output_string(manager.current_step)
    print(conf)
