+++ src/Observables/StructureFactor.h
@@ -56,7 +56,9 @@ public:
 
 	void get_settings(input_file &my_inp, input_file &sim_inp);
 	virtual void init();
-	virtual std::string get_output_string(llint curr_step);
+
+	void update_data(llint curr_step) override;
+	std::string get_output_string(llint curr_step) override;
 };
 
 #endif /* STRUCTUREFACTOR_H_ */
