### 2022-07-08 72ec9fd
updated oat's dependency check to give an error for the cython header problem.  Also changed name of the check function because dependencies is too hard to spell.
