+++ analysis/src/oxDNA_analysis_tools/backbone_flexibility.py
@@ -122,8 +122,8 @@ def main():
     args = parser.parse_args()
 
     #run system checks
-    from oxDNA_analysis_tools.config import check_dependencies
-    check_dependencies(["python", "numpy", "matplotlib"])
+    from oxDNA_analysis_tools.config import check
+    check(["python", "numpy", "matplotlib"])
 
     top_file  = args.topology[0]
     traj_file = args.trajectory[0]
