+++ analysis/src/oxDNA_analysis_tools/config.py
@@ -47,7 +47,7 @@ def check(to_check:List[str]=["python", "numpy", "matplotlib", "Bio", "sklearn",
         from sys import version_info
         ver = '.'.join([str(i) for i in version_info[0:2]])
         print("INFO: Python version: {}".format('.'.join([str(i) for i in version_info[0:3]])), file=stderr)
-        if version_info < (3, 8):
+        if version_info < (3, 9):
             flag = True
             print("WARNING: Some scripts will not run with Python versions earler than 3.8.  You have {}, please update your environment", file=stderr)
 
