+++ docs/source/install.md
@@ -10,7 +10,7 @@ Compiling with CUDA support requires CMake >= 3.5 and a CUDA toolkit >= 9.0. If 
 
 ### Python bindings
 
-Compiling with the python bindings enabled requires a working Python3 installation comprising both binaries and include files. With Debian-derived distro these come with the package `python3-dev`.
+Compiling with the python bindings enabled requires a working Python3 installation comprising both binaries and include files. With Debian-derived distro these come with the package `python3-dev`.  OxDNA Analysis Tools requires Python version 3.9 or newer.  `oxpy` will work with any version of Python 3.
 
 ## Compiling oxDNA
 
@@ -64,6 +64,10 @@ When compiling with the Python bindings enabled CMake will sometimes choose the 
 cmake .. -DPython=ON -DPYTHON_INCLUDE_DIRS=/path/to/python/include/dir -DPYTHON_EXECUTABLE=/path/to/python/binary
 ```
 
+If you are using conda environments, the paths should look something like:  
+Include: `$HOME/anaconda3/envs/py38/include/python3.8`
+Executable: `$HOME/anaconda3/envs/py38/bin/python`
+
 If you get an error regarding the number of bytes in the `numpy.array` header, this happens when the version of Numpy on your system doesn't match the version that pip downloads from PyPi when installing `oxDNA_analysis_tools` with its isolated environment (most commonly because you installed Numpy using Conda which tends to be a few versions behind PyPi). To fix this, either update your version of Numpy or try to install just `OAT` without build isolation:
   
    `python -m pip install ./analysis --no-build-isolation`
@@ -74,3 +78,5 @@ If you get an error regarding the number of bytes in the `numpy.array` header, t
 * `make test_quick` runs longer tests to check that oxDNA works.
 * `make test_oxpy` checks that the Python bindings work.
 * `make test` runs all sets of tests above.
+* `oat config` will check all dependencies for `OAT`
+* In `oxDNA/analysis/tests` there is a shell script, `test.sh` which will test your `OAT` installation.
