### 2022-07-09 fea9f33
Add a new observable to compute the stress autocorrelation
