### 2022-07-12 db27a2d
added cosin angle potential for bonded and non-bonded beads
