### 2022-07-12 b3bec24
added two different constants for cosin angle potentials
