+++ contrib/tostiguerra/src/CGNucleicAcidsInteraction.h
@@ -70,7 +70,8 @@ protected:
 
 	/// three-body flexibility stuff
 	bool _enable_semiflexibility = false;
-	number _semiflexibility_k;
+	number _semiflexibility_k_bonded;
+	number _semiflexibility_k_nonbonded;
 
 	/// patchy stuff
 	number _deltaPatchMon = 0.5;
@@ -91,7 +92,8 @@ protected:
 	number _WCA(BaseParticle *p, BaseParticle *q, bool update_forces);
 	number _sticky(BaseParticle *p, BaseParticle *q, bool update_forces);
 	number _patchy_three_body(BaseParticle *p, PSBond &new_bond, bool update_forces);
-	number _semiflexibility_three_body(BaseParticle *middle, BaseParticle *n1, BaseParticle *n2, bool update_forces);
+	//number _semiflexibility_three_body(BaseParticle *middle, BaseParticle *n1, BaseParticle *n2, bool update_forces);
+	number _semiflexibility_three_body(BaseParticle *p, BaseParticle *q, bool update_forces);
 
 public:
 	enum {
