### 2022-07-14 36788c3
updated get_confs to handle times larger than the max value of int
