+++ analysis/src/oxDNA_analysis_tools/UTILS/data_structures.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 from dataclasses import dataclass
 from typing import List
+from ctypes import c_ulong
 import numpy as np
 
 @dataclass
