### 2022-07-14 9bd69a9
remove k from cosin angle potential for non-bonded beads
