### 2022-07-14 9aaee70
remove k from cosin angle potential for non-bonded beads
