### 2022-07-23 6349703
added cli documentation to oat
