+++ analysis/src/oxDNA_analysis_tools/superimpose.py
@@ -5,12 +5,15 @@ from sys import stderr
 from oxDNA_analysis_tools.UTILS.RyeReader import get_confs, describe, inbox, write_conf
 from oxDNA_analysis_tools.align import svd_align
 
-
-def main():
-    parser = argparse.ArgumentParser(prog = os.path.basename(__file__), description="superimposes one or more structures sharing a topology to a reference structure")
+def cli_parser(prog="superimpose.py"):
+    parser = argparse.ArgumentParser(prog = prog, description="superimposes one or more structures sharing a topology to a reference structure")
     parser.add_argument('reference', type=str, nargs=1, help="The reference configuration to superimpose to")
     parser.add_argument('victims', type=str, nargs='+', help="The configuraitons to superimpose on the reference")
     parser.add_argument('-i', metavar='index_file', dest='index_file', nargs=1, help='Align to only a subset of particles from a space-separated list in the provided file')
+    return parser
+
+def main():
+    parser = cli_parser(os.path.basename(__file__))    
     args = parser.parse_args()
 
     #run system checks
