+++ analysis/src/oxDNA_analysis_tools/db_to_force.py
@@ -38,12 +38,15 @@ def parse_dot_bracket(input:str) -> List[int]:
 
     return output
 
-
-def main():
-    parser = argparse.ArgumentParser(prog = os.path.basename(__file__), description="Create an external forces file enforcing the current base-pairing arrangement")
+def cli_parser(prog="db_to_force.py"):
+    parser = argparse.ArgumentParser(prog = prog, description="Create an external forces file enforcing the current base-pairing arrangement")
     parser.add_argument('db_file', type=str, nargs=1, help="A text file containing dot-bracket notation of the base-pairing arrangement")
     parser.add_argument('-o', '--output', type=str, nargs=1, help='Name of the file to write the force list to')
     parser.add_argument('-s', '--stiffness', type=float, nargs=1, help='Strength of the forces')
+    return parser
+
+def main():
+    parser = cli_parser(os.path.basename(__file__))
     args = parser.parse_args()
 
     from oxDNA_analysis_tools.config import check
