### 2022-07-23 f67857e
removed mappings previously for testing. adding back
