+++ docs/docs_requirements.txt
@@ -3,4 +3,4 @@ sphinx_rtd_theme==0.5.0
 myst_parser==0.17.2
 docutils==0.16
 Jinja2<3.1
-sphinx-argparse>0.3.1
+sphinx-argparse>=0.3.1
