### 2022-07-26 bbac17b
updated get_confs function call to be cleaner
