+++ analysis/src/oxDNA_analysis_tools/contact_map.py
@@ -27,7 +27,7 @@ def compute(ctx:ComputeContext, chunk_size:int,  chunk_id:int) -> np.array:
     Returns:
         distances (numpy.array): A NxN matrix containing pairwise distances between every pair of nucleotides.
     """
-    confs = get_confs(ctx.traj_info.idxs, ctx.traj_info.path, chunk_id*chunk_size, chunk_size, ctx.top_info.nbases)
+    confs = get_confs(ctx.top_info, ctx.traj_info, chunk_id*chunk_size, chunk_size)
 
     np_poses = np.asarray([c.positions for c in confs])
     distances = np.zeros((ctx.top_info.nbases, ctx.top_info.nbases))
