+++ analysis/src/oxDNA_analysis_tools/backbone_flexibility.py
@@ -36,7 +36,7 @@ def compute(ctx:ComputeContext, chunk_size:int, chunk_id:int):
     torsions = np.zeros(ctx.top_info.nbases-(2*len(ctx.system.strands)))
     dihedrals = np.zeros(ctx.top_info.nbases-(3*len(ctx.system.strands)))
 
-    confs = get_confs(ctx.traj_info.idxs, ctx.traj_info.path, chunk_id*chunk_size, chunk_size, ctx.top_info.nbases)
+    confs = get_confs(ctx.top_info, ctx.traj_info, chunk_id*chunk_size, chunk_size)
     for conf in confs:
         for i, strand in enumerate(ctx.system):
             for j, n in enumerate(strand):

[…]
