### 2022-07-26 7101b9a
bump oat version with get_confs change
