### 2022-07-28 00f8328
Added minimal jupyter example to docs and fixed crash from running multiple oxpy simulations from a cell
