+++ docs/source/conf.py
@@ -78,6 +78,7 @@ templates_path = ['_templates']
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = []
+html_extra_path = []
 
 try:
     import oxpy
@@ -95,7 +96,7 @@ html_theme = 'sphinx_rtd_theme'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = []
+html_static_path = ["./images/oxView_frame"]
 
 def setup(app):
     app.add_config_value('with_oxpy', True, 'env')
