+++ _sources/oat/index.md.txt
@@ -59,3 +59,125 @@ Full API documentation:
 
    forces.md
 ```
+
+## Analysis notebooks
+As simulations become more prevalent and analysis pipelines more complicated, it can be beneficial for researchers to define their entire simulation/analysis pipeline in Python notebooks, much as is often done in the machine learning community.  The modular nature of `oxpy` and `oxDNA_analysis_tools` lends itself to composing analyses together to ask specific scientific questions.
+
+This section represents a minimal example of how to run, analyze and visualize a simulation from within a [Jupyter Notebook](https://jupyter.org/).  For a more comprehensive example which includes live plotting of values while the simulation is running, see the [OXPY_JUPYTER](https://github.com/lorenzo-rovigatti/oxDNA/blob/master/examples/OXPY_Jupyter/literate_sim.ipynb) example in the `oxDNA/examples` directory.
+
+From a directory containing an oxDNA input file, and initial configuration/topology files:
+
+**First, we will set up some functions for running simulations with oxpy**
+```python
+import oxpy
+import multiprocessing
+
+# Running oxpy multiple times from the same jupyter kernel crashes the kernel for some reason
+# So we will be starting simulations from a separate thread.
+def spawn(f, kwargs = {}):
+    p = multiprocessing.Process(target=f, kwargs = kwargs)
+    p.start()
+    return p
+
+# Sets up and runs the simulation
+def run(**kwargs):
+    with oxpy.Context():
+        inp = oxpy.InputFile()
+        inp.init_from_filename("input")
+
+        # Modify inputfile at runtime
+        # all input parameters provided to oxDNA need to be strings 
+        for k,v in kwargs.items(): 
+            inp[k] = v
+        
+        manager = oxpy.OxpyManager(inp)
+        manager.run_complete()
+
+p = None
+```
+
+**Next, we will initiate a simulation.**  
+For this example, we set the number of steps and print intervals to be smaller such that this simulation finishes in a reasonable amount of time
+
+```python
+# Kill any currently running simulations
+try:
+    p.terminate()
+except:
+    pass
+
+# You can pass modifications to a default input file as dict of key : value pairs
+# Both key and values must be strings
+input_mods = {
+    "steps" : "1e6",
+    "print_conf_interval" : "1e4",
+    "print_energy_every" : "1e4"
+}
+
+# Run a simulation and obtain a reference to the background process
+p = spawn(run, input_mods)
+
+# The input file used to run the simulation gets lost in the subprocess, 
+# so make it again so we have access to it for analysis
+with oxpy.Context():
+    inp = oxpy.InputFile()
+    inp.init_from_filename("input")
+    for k,v in input_mods.items(): 
+        inp[k] = v
+```
+
+**You can stop a running simulation by terminating its process:**
+```python
+# You can stop the simulation at any time by running this cell.
+p.terminate()
+```
+
+**Use `oat` to compute the mean structure and RMSF**
+```python
+from oxDNA_analysis_tools.mean import mean
+from oxDNA_analysis_tools.deviations import deviations
+from oxDNA_analysis_tools.UTILS.RyeReader import describe
+
+# Get the trajectory and topology file names from the oxpy.InputFile object
+top = inp["topology"]
+traj = inp["trajectory_file"]
+top_info, traj_info = describe(top, traj)
+
+# Compute the mean structure and RMSFs
+mean_conf = mean(traj_info, top_info, ncpus=4)
+RMSDs, RMSFs = deviations(traj_info, top_info, mean_conf, ncpus=4)
+
[…]
