+++ docs/source/oat/index.md
@@ -68,6 +68,7 @@ This section represents a minimal example of how to run, analyze and visualize a
 From a directory containing an oxDNA input file, and initial configuration/topology files:
 
 **First, we will set up some functions for running simulations with oxpy**
+Instead of modifying an existing input file, you can also simply provide the entire input file as a `dict` to kwargs.
 ```python
 import oxpy
 import multiprocessing
@@ -97,7 +98,7 @@ p = None
 ```
 
 **Next, we will initiate a simulation.**  
-For this example, we set the number of steps and print intervals to be smaller such that this simulation finishes in a reasonable amount of time
+For this example, we set the number of steps and print intervals to be smaller such that this simulation finishes in a reasonable amount of time.
 
 ```python
 # Kill any currently running simulations
