### 2022-07-29 733aab1
added potential on a_1 of non-bonded beads
