+++ contrib/tostiguerra/src/CGNucleicAcidsInteraction.cpp
@@ -210,19 +210,22 @@ number CGNucleicAcidsInteraction::_sticky(BaseParticle *p, BaseParticle *q, bool
 			number cost = (p->orientationT.v3 * q->orientationT.v3);
 			number Vteta = (1. - cost)/2.;
 
-			number tmp_energy = Vradial * Vteta;
+			number cost_a1 = (p->orientationT.v1 * q->orientationT.v1);
+			number Vteta_a1 = (1. - cost_a1)/2.;
+
+			number tmp_energy = Vradial * Vteta * Vteta_a1;
 			energy += tmp_energy;
 
-			number tb_energy = (r_mod < _3b_sigma) ? epsilon * Vteta : -tmp_energy;
+			number tb_energy = (r_mod < _3b_sigma) ? epsilon * Vteta * Vteta_a1: -tmp_energy;
 
 			PSBond p_bond(q, tb_energy, epsilon, 1, 1, patch_dist);
 			PSBond q_bond(p, tb_energy, epsilon, 1, 1, -patch_dist);
 
 			if(update_forces) {
-				number force_mod = ( epsilon * _3b_A_part * exp_part * (4. * _3b_B_part / (SQR(sqr_r) * r_mod)) + _3b_sigma * Vradial / SQR(r_mod - _3b_rcut) ) * Vteta;
+				number force_mod = ( epsilon * _3b_A_part * exp_part * (4. * _3b_B_part / (SQR(sqr_r) * r_mod)) + _3b_sigma * Vradial / SQR(r_mod - _3b_rcut) ) * Vteta * Vteta_a1;
 				LR_vector tmp_force = patch_dist * (-force_mod / r_mod);
 
-				LR_vector torque_tetaTerm = Vradial * ( p->orientationT.v3.cross(q->orientationT.v3) )/2.;
+				LR_vector torque_tetaTerm = Vradial * ( p->orientationT.v3.cross(q->orientationT.v3) )/2. * Vteta_a1 + Vradial * ( p->orientationT.v1.cross(q->orientationT.v1) )/2. * Vteta;
 				LR_vector p_torque = p->orientationT * ( p_patch_pos.cross(tmp_force) + torque_tetaTerm);
 				LR_vector q_torque = q->orientationT * ( q_patch_pos.cross(tmp_force) + torque_tetaTerm);
 
