### 2022-07-29 8734553
Make the code deallocate ConfigInfo at the end of each simulation
