### 2022-07-30 2271ccb
Remove a duplicated lib
