### 2022-07-30 25f81c4
Make oxpy's two setup files a bit more uniform with each other
