### 2022-07-30 81fe23c
Remove some IDE warnings and improve an error message
