### 2022-07-30 eedcf0d
Fix a bug whereby oxpy simulations on CUDA couldn't be run consequently
