### 2022-08-03 be0e1e4
Added docstrings for oxView
