+++ analysis/src/oxDNA_analysis_tools/superimpose.py
@@ -1,14 +1,50 @@
 import argparse
 import os
+from typing import List
 import numpy as np
 from sys import stderr
 from oxDNA_analysis_tools.UTILS.RyeReader import get_confs, describe, inbox, write_conf
+from oxDNA_analysis_tools.UTILS.data_structures import Configuration
 from oxDNA_analysis_tools.align import svd_align
+
+
+def superimpose(ref:Configuration, victims:List[Configuration], indexes:List[int]=None):
+    """
+    Superimposes one or more structures sharing a topology to a reference structure
+
+    Parameters:
+        ref (Configuration) : the reference configuration to superimpose to
+        victims (List[Configuration]) : the configurations to superimpose on the reference
+        indexes (List[int]) : the indexes of the particles to superimpose on the reference (default: all)
+
+    Returns:
+        Aligned configurations (List[Configuration])
+    """
+
+    ref = inbox(ref)
+    # alignment requires the ref to be centered at 0.  Inboxing did not take the indexing into account.
+    reference_coords = ref.positions[indexes]
+    ref_cms = np.mean(reference_coords, axis=0) # cms prior to centering
+    reference_coords = reference_coords - ref_cms
+
+    aligned = []
+
+    for i, f in enumerate(victims):
+        top_info, traj_info = describe(None, f)
+        conf = get_confs(top_info, traj_info, 0, 1)[0]
+        conf = inbox(conf)
+
+        np_coords = np.asarray([conf.positions, conf.a1s, conf.a3s])
+
+        conf.positions, conf.a1s, conf.a3s = svd_align(reference_coords, np_coords, indexes)
+        aligned.append(conf)
+
+    return aligned
 
 def cli_parser(prog="superimpose.py"):
     parser = argparse.ArgumentParser(prog = prog, description="superimposes one or more structures sharing a topology to a reference structure")
     parser.add_argument('reference', type=str, nargs=1, help="The reference configuration to superimpose to")
-    parser.add_argument('victims', type=str, nargs='+', help="The configuraitons to superimpose on the reference")
+    parser.add_argument('victims', type=str, nargs='+', help="The configurations to superimpose on the reference")
     parser.add_argument('-i', metavar='index_file', dest='index_file', nargs=1, help='Align to only a subset of particles from a space-separated list in the provided file')
     return parser
 
@@ -40,20 +76,9 @@ def main():
     else: 
         indexes = list(range(top_info.nbases))
 
-    # alignment requires the ref to be centered at 0.  Inboxing did not take the indexing into account.
-    reference_coords = ref_conf.positions[indexes]
-    ref_cms = np.mean(reference_coords, axis=0) # cms prior to centering
-    reference_coords = reference_coords - ref_cms
+    aligned = superimpose(ref_conf, args.victims, indexes)
 
-    for i, f in enumerate(args.victims):
-        top_info, traj_info = describe(None, f)
-        conf = get_confs(top_info, traj_info, 0, 1)[0]
-        conf = inbox(conf)
-
-        np_coords = np.asarray([conf.positions, conf.a1s, conf.a3s])
-
-        conf.positions, conf.a1s, conf.a3s = svd_align(reference_coords, np_coords, indexes)
-
+    for i, conf in enumerate(aligned):
         write_conf("aligned{}.dat".format(i), conf)
         print("INFO: Wrote file aligned{}.dat".format(i), file=stderr)
 
