### 2022-08-04 a0b7921
Improve an error message and update the CUDA requirements
