### 2022-08-04 e86be92
Remove a deprecated function by reworking how textures are accessed
