### 2022-08-12 3769e70
Add documentation for the dt option as per #25
