### 2022-08-12 2a52d35
Resolve the ambiguity about `restart_step_counter` (see #28)
