### 2022-08-12 5e325ac
Fix #27
The code interpreted the `use_average_seq` option as an integer instead
of a boolean (still used `getInputInt` instead of `getInputBool`)
