+++ src/Interactions/DNAInteraction.cpp
@@ -229,17 +229,14 @@ DNAInteraction::~DNAInteraction() {
 void DNAInteraction::get_settings(input_file &inp) {
 	BaseInteraction::get_settings(inp);
 
-	int avg_seq;
-	if(getInputInt(&inp, "use_average_seq", &avg_seq, 0) == KEY_FOUND) {
-		_average = (bool) avg_seq;
-		if(!_average)
+	if(getInputBool(&inp, "use_average_seq", &_average, 0) == KEY_FOUND) {
+		if(!_average) {
 			getInputString(&inp, "seq_dep_file", _seq_filename, 1);
-		OX_LOG(Logger::LOG_INFO, "Using '%s' as the input for sequence-dependent values", _seq_filename.c_str());
+			OX_LOG(Logger::LOG_INFO, "Using '%s' as the input for sequence-dependent values", _seq_filename.c_str());
+		}
 	}
 
-	float hb_multi = 1.;
-	getInputFloat(&inp, "hb_multiplier", &hb_multi, 0);
-	_hb_multiplier = (number) hb_multi;
+	getInputNumber(&inp, "hb_multiplier", &_hb_multiplier, 0);
 
 	char T[256];
 	getInputString(&inp, "T", T, 1);
