### 2022-08-17 #4553 @ppxasjsm
@lorenzo-rovigatti, I am Antonia Mey and I'll be your editor for the submission. I am new to this role as editor for JOSS so please bear with me.

I am now in the process of finding reviewers for you.
