### 2022-08-18 28cedeb
fixed typo in db_to_force parser
