### 2022-08-19 719836c
Fix a bug related to the python bindings of the analysis backend
Also add a test to catch these problems early
