+++ src/Backends/AnalysisBackend.h
@@ -52,8 +52,8 @@ public:
 
 	const FlattenedConfigInfo &flattened_conf();
 
-	ConfigInfo *config_info() {
-		return _config_info;
+	std::shared_ptr<ConfigInfo> config_info() {
+		return CONFIG_INFO;
 	}
 
 	llint get_conf_step() {
