### 2022-08-22 85bccd6
Fix a compilation issue of a contrib observable
