### 2022-08-22 7e437f6
Add methods to sync data from and to CUDA interactions
