### 2022-08-24 3432d41
Update a CUDA error message and fix a mistake in the documentation
