### 2022-08-30 #4693 @ptmerz
Hi @ppxasjsm, I have started playing around with the software, but have not had enough time to dig deeper yet, so haven't generated the checklist yet. I should be fine generating it, though. This week is much calmer, so I should be able to start checking off items soon!
