### 2022-09-01 ec1be2b
Fix a bug where confs without momenta would be printed w trailing spaces
