### 2022-09-07 77717c9
Change libcommon.so to liboxdna_common.so.
