### 2022-09-07 a52d249
Use a forward declaration for the json main class
This should fix a bug whereby the code wouldn't compile with CUDA 10.x
under some circumstainces
