+++ src/Utilities/parse_input/parse_input.cpp
@@ -5,6 +5,7 @@
 #include "../oxDNAException.h"
 #include "../ConfigInfo.h"
 
+#include <nlohmann/json.hpp>
 #include <exprtk/exprtk.hpp>
 
 #include <regex>
