+++ src/Observables/StressAutocorrelation.cpp
@@ -7,6 +7,8 @@
 
 #include "StressAutocorrelation.h"
 
+#include <sstream>
+
 StressAutocorrelation::StressAutocorrelation() :
 				BaseObservable() {
 
