+++ src/Observables/ObservableFactory.cpp
@@ -58,6 +58,8 @@
 #include "Configurations/TEPxyzOutput.h"
 #include "Configurations/JordanOutput.h"
 
+#include <nlohmann/json.hpp>
+
 ObservablePtr ObservableFactory::make_observable(input_file &obs_inp) {
 	char obs_type[512];
 	getInputString(&obs_inp, "type", obs_type, 1);
