### 2022-09-07 7a0c5b8
Make it possible to compile oxpy with -DDebug=True
