### 2022-09-07 e003f6c
Use the same name for the oxDNA common library for all build types
