### 2022-09-11 #4693 @lorenzo-rovigatti
I never thought that oxDNA would be so intimidating to potential reviewers :p
