### 2022-09-15 #4693 @ptmerz
Here's a few points from my first pass, mostly focussing on the paper and more formal aspects of the submission. I will get back with feedback on the program itself and the documentation soon.

In general, my first impression of the oxDNA software and its associated paper is very positive. There is a clearly defined statement of need, and the software and the model is well-established, proving its use to the community, and the paper is clear and well-written. The points below are hence minor suggestions:

### Contributions and authorship
With respect to the following point in the checklist:
> Contribution and authorship: Has the submitting author (@lorenzo-rovigatti) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

There is no doubt that @lorenzo-rovigatti contributed significantly. However, Github lists 8 contributors (https://github.com/lorenzo-rovigatti/oxDNA/graphs/contributors), but the paper lists only 5 authors. It seems that the contributors not listed as authors had either minor contributions, or are no longer active. It all looks reasonable to me, but maybe @lorenzo-rovigatti could comment on how the decision was made, and @ppxasjsm can confirm that this is ok by JOSS rules.

### Paper
Minor suggestions:
* I would have appreciated a short paragraph early in the paper that lists and defines the different concepts named "oxDNA". I think to have understood that it stands for 1. a coarse-grained DNA model, 2. an update to that first model, named oxDNA2, 3. the software package reviewed here, implementing MD and MC simulations using the models of the same name, 4. a LAMMPS pair style implementing the oxDNA and oxDNA2 models. Later in the paper, the authors also refer to other software packages using "oxDNA configurations", which presumably refers to configurations created by the oxDNA program (3.). I found these naming issues challenging for someone (like myself) who is not familiar with these models and the software ecosystem surrounding them, and in fact it might well be that I still misunderstood something!
* I think that the earlier publications about the software [P. Šulc et al., J. Chem. Phys. 137, 135101 (2012), L. Rovigatti et al., J. Comput. Chem. 36, 1 (2015)] should be cited. It would also be good to make it more explicit what the improvements compared to these previous publications is. I assume that the paragraph starting on line 51 (_This new version of the code is now hosted on GitHub and has been modernized ..._) refers to the version published earlier, but I would suggest to make this explicit.
* The paper contains a link to underline the importance of oxDNA, showing that it has been used in over a hundred publications.
  - The link is outdated, pointing to a now defunct publons page. It does get automatically forwarded (at least for now) to its new version on Web of Science, but it would be better to directly link to the new site.
  - It is unclear how the publications for oxDNA / oxRNA have been counted, and whether this count refers to the software package, or also includes citations of the model / usages of the LAMMPS implementation. Also, how does that count relate to the citation count for the earlier papers describing the software?
  - More generally, I was wondering if there were any editorial guidelines (@ppxasjsm?) regarding links in the paper pdf file? Are there any requirements, like a clear-text version of the linked website? The current link destination is only visible when hovering the mouse pointer over the link text.

### Documentation

@lorenzo-rovigatti, could you comment on the relationship between https://lorenzo-rovigatti.github.io/oxDNA/index.html and https://dna.physics.ox.ac.uk/index.php/Main_Page, both of which are linked in different places of your Github repo? Should both be seen as part of the current submission or is one superseding the other?
