### 2022-09-16 #4693 @lorenzo-rovigatti
Hey @ptmerz , thanks a lot for all your comments, I really appreciated the effort you are putting into improving this submission!

Here are some answers to your remarks.

## Authorship

FerdinandoR was a PhD student who wrote a custom interaction (with related analysis classes) to study a specific model that was the focus of his thesis. He did not contribute to the core package, and his last commit dates back to 2018, which predates (by a lot) the current rewrite which is the focus of the submission. debeshmandal is an oxDNA user who helped me setting up CI scripts and cgevans had a merged PR to help himself (and possibly other maintainers) packaging oxDNA for linux distro. Having said all of this, I also think that it makes sense to wait for @ppxasjsm 's opinion on this!

## Paper

We have updated the paper to reflect your comments, which are all very sensible and appreciated! In particular, now the link to the publications refer to the oxDNA family models rather than the software, we have added a mention to the 2015 paper and we added a couple of sentences to warn the reader that, unfortunately, there are several things that are named "oxDNA".

## Documentation

The .oc.ac.uk page contains details on the model itself (rather than the software package). I have contacted the site maintainer and asked them to remove the old pages documenting the software and add links to the new oxDNA repo. I hope the switch will happen soon :-)
