### 2022-09-21 #4693 @ppxasjsm
> More generally, I was wondering if there were any editorial guidelines (@ppxasjsm?) regarding links in the paper pdf file? Are there any requirements, like a clear-text version of the linked website? The current link destination is only visible when hovering the mouse pointer over the link text.

I am not aware of anything specific in the guidelines after a quick search, but making them clearly visible as links would be helpful I think.

> Authorship

FerdinandoR was a PhD student who wrote a custom interaction (with related analysis classes) to study a specific model that was the focus of his thesis. He did not contribute to the core package, and his last commit dates back to 2018, which predates (by a lot) the current rewrite which is the focus of the submission. debeshmandal is an oxDNA user who helped me setting up CI scripts and cgevans had a merged PR to help himself (and possibly other maintainers) packaging oxDNA for linux distro. Having said all of this, I also think that it makes sense to wait for @ppxasjsm 's opinion on this!

The guidelines state the following:

> Authorship
> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are. The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed. In addition, co-authors agree to be accountable for all aspects of the work, and to notify JOSS if any retraction or correction of mistakes are needed after publication.

I think debeshmandal can/should be attributed, as he contributed to making the repository testable and adhering to modern coding standards unless @lorenzo-rovigatti has strong objections for him to be included. For the other two I do not see the need to be named authors based on the guidelines.
