### 2022-09-26 #4693 @degiacom
I find the article is overall well-written. Beyond the pertinent observations by @ptmerz, I have two additional comments.

- it is expected that the Summary section features "_a clear description of the high-level functionality and purpose of the software_". In this spirit, I suggest the authors add one sentence stating what oxDNA is/does.
- In the section "_Related Software_", I find that that the two points the authors make, when comparing the oxDNA software with LAMMPS, could be improved. Specifically:
  - "_the generic and hyper customizable nature of the LAMMPS package does not lend itself well to being used by non specialists_". This statement seems to imply that a non-specialist could run simulations with oxDNA. Perhaps this is a bit of a stretch: while no specific knowledge on usage of HPC infrastructure is required, a user would still need to understand molecular simulation principles, know how to use a terminal, and how to write appropriate input files. I suggest it could be stated that the oxDNA software provides an easier entry point to run simulations with the oxDNA model (or a statement to this effect).
  - "_[...] it lacks many of the tools required to ease the burden of working with coarse grained nucleic acids_". This sounds quite vague, I think it would be useful to explicitly provide one or few examples of what those burdens are, and how the tools within oxDNA ease them.
