### 2022-09-26 #4693 @lorenzo-rovigatti
Thanks for the review! I have updated the paper following your comments. Here is a list of changes:

- Added the following sentence at the end of the Summary section: "Here we present `oxDNA`, a software package that has been designed to efficiently run coarse-grained simulations of DNA and RNA and also features an analysis suite aimed at post-processing the analysis of oxDNA/oxRNA trajectories."
- The paragraph on the LAMMPS-oxDNA comparison has been rewritten, and the two sentences mentioned by @degiacom have been replaced by "However, at the moment it lacks GPU support, it cannot be used to build free-energy profiles by biasing discrete reaction coordinates (*e.g.* number of correct basepairs) and there are no tools to analyse LAMMPS-generated trajectories such as `oat` [@oat] or `oxview` [@bohlin2022design]. In addition, the LAMMPS engine supports many more (often oxDNA-unrelated) options, which may make it harder to correctly setup oxDNA simulations compared to `oxDNA`."
- I have also updated the acknowledgements sections as requested by @debeshmandal
