+++ analysis/src/oxDNA_analysis_tools/oxDNA_PDB2.py
@@ -4,31 +4,79 @@ import sys
 import os
 import numpy as np
 import copy
-import string
 import argparse
-
-import oxpy
+import string
+from collections import defaultdict
+from math import sqrt, sin
 
 from oxDNA_analysis_tools.UTILS.pdb import Atom, Nucleotide, AminoAcid, FROM_OXDNA_TO_ANGSTROM
-from oxDNA_analysis_tools.UTILS.RyeReader import get_confs, describe, strand_describe, get_input_parameter
+from oxDNA_analysis_tools.UTILS.RyeReader import get_confs, describe, strand_describe, get_input_parameter, inbox
 import oxDNA_analysis_tools.UTILS.protein_to_pdb as pro
 import oxDNA_analysis_tools.UTILS.utils as utils
 
 DD12_PDB_PATH = "./UTILS/dd12_na.pdb"
 
-def main():
-    print("This script isn't working yet.  Check back later.")
-    sys.exit()
+number_to_base = {0 : 'A', 1 : 'G', 2 : 'C', 3 : 'T'}
 
-    parser = argparse.ArgumentParser(prog = os.path.basename(__file__), description="Convert oxDNA files to PDB.  This converter can handle oxDNANM protein simulation files.")
-    parser.add_argument('input', type=str, nargs=1, help='the input file used to run the oxDNA simulation')
-    parser.add_argument('configuration', type=str, nargs=1, help='the configuration file you wish to convert')
-    parser.add_argument('direction', type=str, nargs=1, help='the direction of strands in the oxDNA files, either 35 or 53.  Most oxDNA files are 3-5.')
-    parser.add_argument('pdbfiles', type=str, nargs='?', help='PDB files for the proteins present in your structure.  If you have multiple proteins, you must specify multiple PDB files.')
-    parser.add_argument('-H', '--hydrogen', action='store_true', default=True, help='if you want to include hydrogen atoms in the output PDB file')
-    parser.add_argument('-u', '--uniform-residue-names', action='store_true', default=False, help='if you want to use uniform residue names in the output PDB file')
-    parser.add_argument('-o', '--one_file_per_strand', action='store_true', default=False, help='if you want to have one PDB file per strand')
-    parser.add_argument('-s', '--same-pdb-all-protein-strands', action='store_true', default=False, help='if you want to have the same PDB file for all protein strands')
+
+base_to_number = {'A' : 0, 'a' : 0, 'G' : 1, 'g' : 1,
+                  'C' : 2, 'c' : 2, 'T' : 3, 't' : 3,
+                  'U' : 3, 'u' : 3, 'D' : 4}
+                  
+aa_to_number = {'A':-1, 'R':-2, 'N':-3, 'D':-4, 'C':-5, 
+                'E':-6, 'Q':-7, 'G':-8, 'H':-9, 'I':-10, 
+                'L':-11, 'K':-12, 'M':-13, 'F':-14, 
+                'P':-15, 'S':-16, 'T':-17, 'W':-18, 
+                'Y':-19, 'V':-20, 'Z':-21, 'X':0}
+
+number_to_aa = {-1:'A', -2:'R', -3:'N', -4:'D', -5:'C', 
+                -6:'E', -7:'Q', -8:'G', -9:'H', -10:'I', 
+                -11:'L', -12:'K', -13:'M', -14:'F', 
+                -15:'P', -16:'S', -17:'T', -18:'W', 
+                -19:'Y', -20:'V', -21:'Z', 0:'X'}
+
+def align(full_base, ox_base):
+        theta = utils.get_angle(full_base.a3, ox_base._a3)
+        # if the two bases are already essentially aligned then we do nothing
[…]

[…]
