### 2022-10-01 686fb30
oxDNA_PDB2 now works up to the mysterious last section
