+++ analysis/src/oxDNA_analysis_tools/UTILS/pdb.py
@@ -61,23 +61,21 @@ class Atom(object):
         r = 0.5
         return "%s %s %s @ %f C[%s]" % (self.pos[0], self.pos[1], self.pos[2], r, color)
 
-class Nucleotide(object):
+class Nucleotide():
     RNA_warning_printed = False
     
     def __init__(self, name, idx):
-        object.__init__(self)
         self.name = name.strip()
         if self.name in NAME_TO_BASE.keys():
             self.base = NAME_TO_BASE[self.name]
         elif self.name in BASES:
             if self.name == "U" and not Nucleotide.RNA_warning_printed:
-                print >> sys.stderr, "WARNING: unsupported uracil detected: use at your own risk"
+                print("WARNING: unsupported uracil detected: use at your own risk", file=sys.stderr)
                 Nucleotide.RNA_warning_printed = True
                 
             self.base = self.name
         else:
             self.base = name[1:]
-        self.atoms:List[Atom] = []
         self.idx = idx
         self.base_atoms = []
         self.phosphate_atoms = []
