+++ analysis/src/oxDNA_analysis_tools/oxDNA_PDB2.py
@@ -36,22 +36,22 @@ number_to_aa = {-1:'A', -2:'R', -3:'N', -4:'D', -5:'C',
                 -19:'Y', -20:'V', -21:'Z', 0:'X'}
 
 def align(full_base, ox_base):
-        theta = utils.get_angle(full_base.a3, ox_base._a3)
+        theta = utils.get_angle(full_base.a3, ox_base['a3'])
         # if the two bases are already essentially aligned then we do nothing
         if sin(theta) > 1e-3:
-            axis = np.cross(full_base.a3, ox_base._a3)
+            axis = np.cross(full_base.a3, ox_base['a3'])
             axis /= sqrt(np.dot(axis, axis))
             R = utils.get_rotation_matrix(axis, theta)
             full_base.rotate(R)
     
-        theta = utils.get_angle(full_base.a1, ox_base._a1)
+        theta = utils.get_angle(full_base.a1, ox_base['a1'])
         if sin(theta) > 1e-3:
-            axis = np.cross(full_base.a1, ox_base._a1)
+            axis = np.cross(full_base.a1, ox_base['a1'])
             axis /= sqrt(np.dot(axis, axis))
             R = utils.get_rotation_matrix(axis, theta)
             full_base.rotate(R)
 
-def cli_parser():
+def cli_parser(prog="oxDNA_PDB.py"):
     parser = argparse.ArgumentParser(prog=os.path.basename(
         __file__), description="Convert oxDNA files to PDB.  This converter can handle oxDNANM protein simulation files.")
     parser.add_argument('input', type=str, nargs=1,
@@ -98,10 +98,7 @@ def main():
     one_file_per_strand = args.one_file_per_strand
     same_pdb_all_protein_strands = args.same_pdb_all_protein_strands
 
-    top_info, traj_info = describe(top_file, conf_file)
-    system, monomers = strand_describe(top_file)
-
-    # Open PDB File of nice lookin duplex
+    # Open PDB File of nice lookin duplex to get base structures from
     with open(os.path.join(os.path.dirname(__file__), DD12_PDB_PATH)) as f:
         nucleotides = []
         aminoacids = []
@@ -115,6 +112,7 @@ def main():
                     old_residue = na.residue_idx
                 nn.add_atom(na)
 
+    # Create reference oxDNA orientation vectors for the PDB base structures
     bases = {}
     for n in nucleotides:
         n.compute_as()
@@ -127,10 +125,14 @@ def main():
     for n in nucleotides:
         n.a1, n.a2, n.a3 = utils.get_orthonormalized_base(n.a1, n.a2, n.a3)
 
+    # Read oxDNA configuration
     system, elements = strand_describe(top_file)
     ti, di = describe(top_file, conf_file)
     conf = get_confs(ti, di, 0, 1)[0]
+    conf = inbox(conf, center=True)
+    box_angstrom = conf.box * FROM_OXDNA_TO_ANGSTROM
 
+    # Handle RMSF -> bFactor conversion
     rmsf_file = args.rmsf_bfactor
     if rmsf_file:
         with open(rmsf_file) as f:
@@ -148,11 +150,8 @@ def main():
     else:
         rmsf_per_nucleotide = defaultdict(lambda: 1.00)
 
-    ox_nucleotides = []
-    
-    conf = inbox(conf, center=True)
-    box_angstrom = conf.box * FROM_OXDNA_TO_ANGSTROM
-
+    # Process optional conditionals
+    correct_for_large_boxes = False
     if np.any(box_angstrom[box_angstrom > 999]):
         print("At least one of the box sizes is larger than 999: all the atoms which are outside of the box will be brought back through periodic boundary conditions", file=sys.stderr)
         correct_for_large_boxes = True
[…]
