### 2022-10-01 104b0a1
Small Update
Changed output PDB name from TM2.pdb to the topology file name. Deleted repetitive line of code.
