### 2022-10-03 2e71313
corrected the r_cut in sticky interaction
