### 2022-10-05 f50aeec
update gennerate-sa.py rename otfile
