+++ contrib/tostiguerra/src/CGNucleicAcidsInteraction.cpp
@@ -59,7 +59,7 @@ void CGNucleicAcidsInteraction::init() {
 	_3b_B_part = B_ss * pow(_3b_sigma, 4.);
 	_3b_prefactor = _3b_lambda;
 
-	_rcut = _3b_rcut + 2 * _deltaPatchMon;
+	_rcut = _3b_rcut + 2 * (_deltaPatchMon);
 
 	if(_PS_alpha > 0.) {
 		if(_rfene > _rcut) {
@@ -195,13 +195,13 @@ number CGNucleicAcidsInteraction::_sticky(BaseParticle *p, BaseParticle *q, bool
 	LR_vector p_patch_pos = p->int_centers[0];
 	LR_vector q_patch_pos = q->int_centers[0];
 	LR_vector patch_dist = _computed_r + q_patch_pos - p_patch_pos;
-	number sqr_computed_r = _computed_r.norm();
 	number sqr_r = patch_dist.norm();
 	number energy = 0.;
 
-	// sticky-sticky
+	// sticky-sticky 
 	if(_sticky_interaction(p->btype, q->btype)) {
-		if(sqr_computed_r  < _sqr_3b_rcut) {
+		if(sqr_r  < _sqr_3b_rcut) {
+			//printf("entro in sticky\n");
 			number epsilon = _3b_epsilon[p->btype * _interaction_matrix_size + q->btype];
 
 			number r_mod = sqrt(sqr_r);
@@ -210,23 +210,26 @@ number CGNucleicAcidsInteraction::_sticky(BaseParticle *p, BaseParticle *q, bool
 
 			number cost = (p->orientationT.v3 * q->orientationT.v3);
 			number Vteta = (1. - cost)/2.;
+			// Vteta = 1;
 
 			number cost_a1 = (p->orientationT.v1 * q->orientationT.v1);
 			number Vteta_a1 = (1. - cost_a1)/2.;
+			// Vteta_a1 = 1;
 
 			number tmp_energy = Vradial * Vteta * Vteta_a1;
 			energy += tmp_energy;
 
-			number tb_energy = (r_mod < _3b_sigma) ? epsilon * Vteta * Vteta_a1: -tmp_energy;
+			number tb_energy = (r_mod < _3b_sigma) ? Vteta * Vteta_a1: -tmp_energy / epsilon;
 
-			PSBond p_bond(q, tb_energy, epsilon, 1, 1, patch_dist);
-			PSBond q_bond(p, tb_energy, epsilon, 1, 1, -patch_dist);
+			PSBond p_bond(q, tb_energy, epsilon, 1, 1, patch_dist, r_mod);
+			PSBond q_bond(p, tb_energy, epsilon, 1, 1, -patch_dist, r_mod);
 
 			if(update_forces) {
 				number force_mod = ( epsilon * _3b_A_part * exp_part * (4. * _3b_B_part / (SQR(sqr_r) * r_mod)) + _3b_sigma * Vradial / SQR(r_mod - _3b_rcut) ) * Vteta * Vteta_a1;
 				LR_vector tmp_force = patch_dist * (-force_mod / r_mod);
 
 				LR_vector torque_tetaTerm = Vradial * ( p->orientationT.v3.cross(q->orientationT.v3) )/2. * Vteta_a1 + Vradial * ( p->orientationT.v1.cross(q->orientationT.v1) )/2. * Vteta;
+				// torque_tetaTerm = LR_vector();
 				LR_vector p_torque = p->orientationT * ( p_patch_pos.cross(tmp_force) + torque_tetaTerm);
 				LR_vector q_torque = q->orientationT * ( q_patch_pos.cross(tmp_force) + torque_tetaTerm);
 
[…]
