+++ utils/make_testcase.sh
@@ -1,0 +1,2 @@
+FILENAME=$(python3 generate_random_seq.py)
+python3 generate-sa.py 10 $FILENAME
