### 2022-10-16 #4693 @ppxasjsm
I think @degiacom and @ptmerz need to still test some of the software. How are you both getting on with your reviewing tasks?
