+++ contrib/tostiguerra/src/CGNucleicAcidsInteraction.cpp
@@ -198,10 +198,9 @@ number CGNucleicAcidsInteraction::_sticky(BaseParticle *p, BaseParticle *q, bool
 	number sqr_r = patch_dist.norm();
 	number energy = 0.;
 
-	// sticky-sticky 
+	// sticky-sticky
 	if(_sticky_interaction(p->btype, q->btype)) {
 		if(sqr_r  < _sqr_3b_rcut) {
-			//printf("entro in sticky\n");
 			number epsilon = _3b_epsilon[p->btype * _interaction_matrix_size + q->btype];
 
 			number r_mod = sqrt(sqr_r);
@@ -220,15 +219,17 @@ number CGNucleicAcidsInteraction::_sticky(BaseParticle *p, BaseParticle *q, bool
 			energy += tmp_energy;
 
 			number tb_energy = (r_mod < _3b_sigma) ? Vteta * Vteta_a1: -tmp_energy / epsilon;
+			// number tb_energy = (r_mod < _3b_sigma) ? 1 : -Vradial / epsilon;
 
-			PSBond p_bond(q, tb_energy, epsilon, 1, 1, patch_dist, r_mod);
-			PSBond q_bond(p, tb_energy, epsilon, 1, 1, -patch_dist, r_mod);
+			PSBond p_bond(q, tb_energy, epsilon, 1, 1, patch_dist, r_mod, _computed_r);
+			PSBond q_bond(p, tb_energy, epsilon, 1, 1, -patch_dist, r_mod, -_computed_r);
 
 			if(update_forces) {
 				number force_mod = ( epsilon * _3b_A_part * exp_part * (4. * _3b_B_part / (SQR(sqr_r) * r_mod)) + _3b_sigma * Vradial / SQR(r_mod - _3b_rcut) ) * Vteta * Vteta_a1;
 				LR_vector tmp_force = patch_dist * (-force_mod / r_mod);
 
 				LR_vector torque_tetaTerm = Vradial * ( p->orientationT.v3.cross(q->orientationT.v3) )/2. * Vteta_a1 + Vradial * ( p->orientationT.v1.cross(q->orientationT.v1) )/2. * Vteta;
+				//LR_vector torque_tetaTerm = Vradial * ( p->orientationT.v1.cross(q->orientationT.v1) )/2. * Vteta;
 				// torque_tetaTerm = LR_vector();
 				LR_vector p_torque = p->orientationT * ( p_patch_pos.cross(tmp_force) + torque_tetaTerm);
 				LR_vector q_torque = q->orientationT * ( q_patch_pos.cross(tmp_force) + torque_tetaTerm);
@@ -238,6 +239,28 @@ number CGNucleicAcidsInteraction::_sticky(BaseParticle *p, BaseParticle *q, bool
 
 				p->torque += p_torque;
 				q->torque -= q_torque;
+
+				// LR_vector check = _computed_r.cross(-tmp_force) + (p_patch_pos.cross(tmp_force) + torque_tetaTerm) - (q_patch_pos.cross(tmp_force) + torque_tetaTerm);
+				// printf("%lf\n", check.module());
+
+				// if(r_mod > _3b_sigma) {
+				// 		p_bond.force = tmp_force / (epsilon * Vteta * Vteta_a1);
+				// 		p_bond.p_torque = p->orientationT * p_patch_pos.cross(tmp_force) / (epsilon * Vteta * Vteta_a1);
+				// 		p_bond.q_torque = q->orientationT * q_patch_pos.cross(tmp_force) / (epsilon * Vteta * Vteta_a1);
+				//
+				// 		q_bond.force = -tmp_force / (epsilon * Vteta * Vteta_a1);
[…]
