### 2022-10-17 d1a1a45
corrected torques in 3b potential
