+++ contrib/tostiguerra/src/CGNucleicAcidsInteraction.cpp
@@ -219,7 +219,6 @@ number CGNucleicAcidsInteraction::_sticky(BaseParticle *p, BaseParticle *q, bool
 			energy += tmp_energy;
 
 			number tb_energy = (r_mod < _3b_sigma) ? Vteta * Vteta_a1: -tmp_energy / epsilon;
-			// number tb_energy = (r_mod < _3b_sigma) ? 1 : -Vradial / epsilon;
 
 			PSBond p_bond(q, tb_energy, epsilon, 1, 1, patch_dist, r_mod, _computed_r);
 			PSBond q_bond(p, tb_energy, epsilon, 1, 1, -patch_dist, r_mod, -_computed_r);
@@ -229,7 +228,6 @@ number CGNucleicAcidsInteraction::_sticky(BaseParticle *p, BaseParticle *q, bool
 				LR_vector tmp_force = patch_dist * (-force_mod / r_mod);
 
 				LR_vector torque_tetaTerm = Vradial * ( p->orientationT.v3.cross(q->orientationT.v3) )/2. * Vteta_a1 + Vradial * ( p->orientationT.v1.cross(q->orientationT.v1) )/2. * Vteta;
-				//LR_vector torque_tetaTerm = Vradial * ( p->orientationT.v1.cross(q->orientationT.v1) )/2. * Vteta;
 				// torque_tetaTerm = LR_vector();
 				LR_vector p_torque = p->orientationT * ( p_patch_pos.cross(tmp_force) + torque_tetaTerm);
 				LR_vector q_torque = q->orientationT * ( q_patch_pos.cross(tmp_force) + torque_tetaTerm);
@@ -239,28 +237,6 @@ number CGNucleicAcidsInteraction::_sticky(BaseParticle *p, BaseParticle *q, bool
 
 				p->torque += p_torque;
 				q->torque -= q_torque;
-
-				// LR_vector check = _computed_r.cross(-tmp_force) + (p_patch_pos.cross(tmp_force) + torque_tetaTerm) - (q_patch_pos.cross(tmp_force) + torque_tetaTerm);
-				// printf("%lf\n", check.module());
-
-				// if(r_mod > _3b_sigma) {
-				// 		p_bond.force = tmp_force / (epsilon * Vteta * Vteta_a1);
-				// 		p_bond.p_torque = p->orientationT * p_patch_pos.cross(tmp_force) / (epsilon * Vteta * Vteta_a1);
-				// 		p_bond.q_torque = q->orientationT * q_patch_pos.cross(tmp_force) / (epsilon * Vteta * Vteta_a1);
-				//
-				// 		q_bond.force = -tmp_force / (epsilon * Vteta * Vteta_a1);
-				// 		q_bond.p_torque = -q->orientationT * q_patch_pos.cross(tmp_force) / (epsilon * Vteta * Vteta_a1);
-				// 		q_bond.q_torque = -p->orientationT * p_patch_pos.cross(tmp_force) / (epsilon * Vteta * Vteta_a1);
-				// }
-				// else {
-				// 	p_bond.force = LR_vector();
-				// 	p_bond.p_torque = LR_vector();
-				// 	p_bond.q_torque = LR_vector();
-				//
-				// 	q_bond.force = LR_vector();
-				// 	q_bond.p_torque = LR_vector();
-				// 	q_bond.q_torque = LR_vector();
-				// }
 
 				if(r_mod > _3b_sigma) {
 						p_bond.force = tmp_force / epsilon;
@@ -280,15 +256,7 @@ number CGNucleicAcidsInteraction::_sticky(BaseParticle *p, BaseParticle *q, bool
 					q_bond.p_torque = q->orientationT * torque_tetaTerm / Vradial;
[…]
