+++ src/Interactions/DNA3Interaction.cpp
@@ -3,7 +3,12 @@
 #include "../Particles/DNANucleotide.h"
 #include <iostream>
 
+DNA3Interaction::DNA3Interaction() :
+                                DNA2Interaction() {
+                std::cout << "\n \n \n  using DNA3Interaction \n \n \n" ;
+                _grooving = 2; 
+}
 
+DNA3Interaction::~DNA3Interaction() {
 
-
-
+}
