### 2022-10-18 950fa2a
Revert "Updated OAT python binding"
This reverts commit d276b46bfb91892020023bfea05d5df7d6bf924a.
