### 2022-10-18 d92d754
Fix a bug related to observables supporting update_every
Observables supporting update_every do not work if update_every is not
set. This commit should fix it, but a more thorough test is needed.
