+++ src/Observables/ObservableOutput.cpp
@@ -180,6 +180,10 @@ void ObservableOutput::print_output(llint step) {
 	stringstream ss;
 	for(auto it = _obss.begin(); it != _obss.end(); it++) {
 		if(it != _obss.begin()) ss << " ";
+
+		if(!(*it)->is_update_every_set()) {
+			(*it)->update_data(step);
+		}
 		ss << (*it)->get_output_string(step);
 	}
 
