+++ src/Observables/Rdf.cpp
@@ -8,7 +8,6 @@
 #include "Rdf.h"
 
 Rdf::Rdf() {
-	_n_confs = 0;
 	_n_pairs = 0;
 	_nbins = -1;
 	_bin_size = (number) -1.;
@@ -35,9 +34,8 @@ void Rdf::update_data(llint curr_step) {
 		box_side = sides[2];
 	}
 
-	_n_confs += 1;
-
-	if(_max_value > box_side / 2. && _n_confs == 1) OX_LOG(Logger::LOG_WARNING, "Observable Rdf: computing profile with max_value > box_size/2. (%g > %g/2.)", _max_value, box_side);
+	_times_updated++;
+	if(_max_value > box_side / 2. && _times_updated == 1) OX_LOG(Logger::LOG_WARNING, "Observable Rdf: computing profile with max_value > box_size/2. (%g > %g/2.)", _max_value, box_side);
 
 	_n_pairs = 0;
 	for(int i = 0; i < N; i++) {
@@ -63,7 +61,7 @@ std::string Rdf::get_output_string(llint curr_step) {
 	std::stringstream ret;
 	ret.precision(9);
 	double myx = _bin_size / 2.;
-	double norm_factor = 4 * M_PI * _n_confs * _n_pairs * _bin_size / _config_info->box->V();
+	double norm_factor = 4 * M_PI * _times_updated * _n_pairs * _bin_size / _config_info->box->V();
 	for(auto value: _profile) {
 		ret << myx << " " << value / (norm_factor * myx * myx) << std::endl;
 		myx += _bin_size;
