### 2022-10-19 d0da8ef
Change the way the CUDA kernels that compute forces are called
This commit tries to avoid the ugly dynamic cast that was used to detect
the type of CUDA list used to compute the forces at runtime
