+++ src/CUDA/Interactions/CUDARNAInteraction.h
@@ -37,8 +37,8 @@ public:
 	CUDARNAInteraction();
 	virtual ~CUDARNAInteraction();
 
-	void get_settings(input_file &inp);
-	void cuda_init(c_number box_side, int N);
+	void get_settings(input_file &inp) override;
+	void cuda_init(int N) override;
 	c_number get_cuda_rcut() {
 		return this->get_rcut();
 	}

[…]
