### 2022-10-23 ce6cff6
cleanup typing in forces2pairs
