+++ contrib/tostiguerra/src/ParticleFTG.h
@@ -79,21 +79,24 @@ ParticleFTG::ParticleFTG(int N_patches, int nt, number sigma, number deltaPM) :
 	type = btype = nt;
 	int_centers.resize(N_patches);
 	_base_patches.resize(N_patches);
+	_base_patches[0] = LR_vector(1, 0, 0);
+	_base_patches[0].normalize();
+	_base_patches[0] *= deltaPM;
 
-	switch(N_int_centers()) {
-	case 0:
-		break;
-	case 1: {
-		_base_patches[0] = LR_vector(1, 0, 0);
-		break;
-	}
-  default:
-		throw oxDNAException("Unsupported number of patches %d\n", N_int_centers());
-	}
-  for(uint i = 0; i < N_int_centers(); i++) {
-    _base_patches[i].normalize();
-    _base_patches[i] *= deltaPM;
-  }
+	// switch(N_int_centers()) {
+	// case 0:
+	// 	break;
+	// case 1: {
+	// 	_base_patches[0] = LR_vector(1, 0, 0);
+	// 	break;
+	// }
+  // default:
+	// 	throw oxDNAException("Unsupported number of patches %d\n", N_int_centers());
+	// }
+  // for(uint i = 0; i < N_int_centers(); i++) {
+  //   _base_patches[i].normalize();
+  //   _base_patches[i] *= deltaPM;
+  // }
 }
 
 void ParticleFTG::set_positions() {
