### 2022-10-25 852da0d
modified the V_a1 potential
