+++ analysis/src/oxDNA_analysis_tools/output_bonds.py
@@ -1,3 +1,4 @@
+from typing import List
 import numpy as np
 import argparse
 from os import path
@@ -5,14 +6,39 @@ from sys import stderr
 from collections import namedtuple
 from oxDNA_analysis_tools.UTILS.data_structures import TopInfo, TrajInfo
 from oxDNA_analysis_tools.UTILS.oat_multiprocesser import oat_multiprocesser
-from oxDNA_analysis_tools.UTILS.RyeReader import describe
+from oxDNA_analysis_tools.UTILS.RyeReader import describe, get_input_parameter
 import oxpy
 
 ComputeContext = namedtuple("ComputeContext",["traj_info",
                                               "top_info",
                                               "input_file",
                                               "visualize",
-                                              "conversion_factor"])
+                                              "conversion_factor",
+                                              "n_potentials"])
+
+def parse_header(e_txt:str) -> List[str]:
+    e_txt = e_txt.strip('#') # strip leading #
+    e_txt = e_txt.split(',')[0] # remove time section
+    e_list = e_txt.split(' ')[2:] # The first two are id1 and id2 (unless somebody writes a non-pairwise potential, then it's your problem)
+    return e_list 
+
+def get_potentials(ctx) -> List[str]:
+    with oxpy.Context():
+        inp = oxpy.InputFile()
+        inp.init_from_filename(ctx.input_file)
+        inp["list_type"] = "cells"
+        inp["trajectory_file"] = ctx.traj_info.path
+        inp["analysis_bytes_to_skip"] = str(0)
+        inp["confs_to_analyse"] = str(1)
+        inp["analysis_data_output_1"] = '{ \n name = stdout \n print_every = 1e10 \n col_1 = { \n id = my_obs \n type = pair_energy \n } \n }'
+
+        backend = oxpy.analysis.AnalysisBackend(inp)
+
+        backend.read_next_configuration()
+        e_txt = backend.config_info().get_observable_by_id("my_obs").get_output_string(0).strip().split('\n')
+        pot_names = parse_header(e_txt[0])
+
+    return pot_names
 
 def compute(ctx:ComputeContext, chunk_size:int, chunk_id:int):
     with oxpy.Context():
@@ -29,7 +55,7 @@ def compute(ctx:ComputeContext, chunk_size:int, chunk_id:int):
 
         backend = oxpy.analysis.AnalysisBackend(inp)
 
-        # The 8 energies are:
+        # The 9 energies in oxDNA2 are:
         # 0 fene
         # 1 bexc
         # 2 stack
@@ -37,13 +63,14 @@ def compute(ctx:ComputeContext, chunk_size:int, chunk_id:int):
         # 4 hb
         # 5 cr_stack
         # 6 cx_stack
-        # 9 Debye-Huckel
-        # 7 total
+        # 7 Debye-Huckel <- this one is missing in oxDNA1
+        # 8 total
         if ctx.visualize:
-            energies = np.zeros((ctx.top_info.nbases, 9))
+            energies = np.zeros((ctx.top_info.nbases, ctx.n_potentials))
 
         while backend.read_next_configuration():
             e_txt = backend.config_info().get_observable_by_id("my_obs").get_output_string(0).strip().split('\n')
+
             if ctx.visualize:
                 for e in e_txt[1:]:
                     if not e[0] == '#':
@@ -61,9 +88,12 @@ def compute(ctx:ComputeContext, chunk_size:int, chunk_id:int):
                         p = int(e[0])
                         q = int(e[1])
                         l = np.array([float(x) for x in e[2:]])*ctx.conversion_factor
-                        print("{} {} {} {} {} {} {} {} {} {} {}".format(p, q, l[0], l[1], l[2], l[3], l[4], l[5], l[6], l[7], l[8]))
[…]
