### 2022-10-25 7e696fc
resolve merge conflict in output_bonds
