+++ analysis/src/oxDNA_analysis_tools/output_bonds.py
@@ -69,12 +69,7 @@ def compute(ctx:ComputeContext, chunk_size:int, chunk_id:int):
             energies = np.zeros((ctx.top_info.nbases, ctx.n_potentials))
 
         while backend.read_next_configuration():
-<<<<<<< HEAD
-            e_txt = backend.config_info().get_observable_by_id("my_obs").get_output_string(0).strip().split('\n')
-
-=======
             e_txt = backend.config_info().get_observable_by_id("my_obs").get_output_string(backend.config_info().current_step).strip().split('\n')
->>>>>>> 25d1289da35dec23488711e9a30179f8b24e2b31
             if ctx.visualize:
                 for e in e_txt[1:]:
                     if not e[0] == '#':
