### 2022-11-15 eaadcab
added 3body semiflexibility
