### 2022-11-15 40e57b4
added 3body semiflexibility
