### 2022-11-15 93bdcf0
corrected constant in semiflexibility_3b_pot
