### 2022-11-15 64b4816
corrected get input option
