+++ utils/oxdna_to_internal.py
@@ -1,0 +1,267 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Tue Nov 15 14:07:04 2022
+
+@author: andrea bonato
+
+code to map oxdna coordinates (center of mass + base vectors)
+to internal coordinates (slide, twist, etc.)
+
+it takes a oxdna trajectory and topology as input 
+and outputs a trajectory with internal coordinates
+
+
+"""
+import numpy as np
+import math
+from scipy import linalg
+
+###############################
+#Parameteres to map oxdna coordinates to interaction centers
+STACK_X = 0.34
+HB_X = 0.4
+BB_X = -0.34
+BB_Y = -0.34
+
+###############################
+#parameter to map nucleotide centers to Euler translations
+OX_TZU_DISP = [0.0,0.0,0.0]
+
+###############################
+#class for oxdna coordinates
+#c (center),bv (base-backbone vector),bnv (base normal vector)
+#and normal (bv x bnv) are np arrays
+#int_centers returns the positions of the interaction centers
+class oxdna_frame :
+    def __init__(self, c,bv,n):
+        self.center = c
+        self.base_v = -bv #minus sign for Tsukuba convention
+        self.normal = n
+        self.base_norv = np.cross(bv,n)
+    def int_centers(self):
+        back_bone = self.center-BB_X*self.base_v-BB_Y*self.base_norv
+        hbond = self.center+HB_X*self.base_v
+        stack = self.center+STACK_X*self.base_v
+        return(back_bone,hbond,stack)
+    
+##############################  
+#class for euler coordinates
+#pos (translation) is a np array, orientation is a np matrix   
+class eframe :
+    def __init__(self, p,ori):
+        self.pos = p
+        self.orientation = ori
+        
+##############################  
+#class for interanl (intra or inter) coordinates
+#tran (translations) and rot (rotations) are np array   
+class int_coord :
+    def __init__(self, tr,rot):
+        self.tran = tr
+        self.rot = rot
+ 
+##############################
+#base class i.e. 1 nucleotide and its frames
+#members are oxframe, which is an oxdna_frame object
+#and frame, which is a eframe object
+class base :
+    def __init__(self,oxc):
+        self.oxframe = oxc
+        
+        #map oxdna coordinates to euler coordinates
+        p = oxc.center - OX_TZU_DISP
+        """
+        A = np.zeros((3,3), dtype=float)
+        
+        #Euler angles
+        #there's a minus sign because we want the inverse rotation (from rigid body to lab)
+        alpha = -math.atan2(oxc.normal[0],-oxc.normal[1])
+        betha = -math.acos(oxc.normal[2])
+        gamma = -math.atan2(oxc.base_norv[2], oxc.base_v[2])
+        
+        #adeg = alpha/math.pi*180
+        #bdeg = betha/math.pi*180
+        #gdeg = gamma/math.pi*180
+        #print(adeg,bdeg,gdeg)
+
+        #Rotation matrix
+        A[0][0] = math.cos(gamma)*math.cos(alpha)-math.cos(betha)*math.sin(gamma)*math.sin(alpha)
+        A[0][1] = -math.cos(gamma)*math.sin(alpha)-math.cos(betha)*math.cos(alpha)*math.sin(gamma)
+        A[0][2] = math.sin(gamma)*math.sin(betha)
+        
+        A[1][0] = math.cos(alpha)*math.sin(gamma)+math.cos(gamma)*math.cos(betha)*math.sin(alpha)
[…]
