### 2022-11-17 476d2eb
average over trajectory
