+++ utils/oxdna_to_internal.py
@@ -19,10 +19,11 @@ from scipy import linalg
 
 ###############################
 #Parameteres to map oxdna coordinates to interaction centers
-STACK_X = 0.34
-HB_X = 0.4
-BB_X = -0.34
-BB_Y = -0.34
+#A = 0; G = 1, C = 2, T = 3
+STACK_X = [0.34, 0.34, 0.34, 0.34]
+HB_X = [0.4, 0.4, 0.4, 0.4]
+BB_X = [-0.34, -0.34, -0.34, -0.34]
+BB_Y = [-0.34, -0.34, -0.34, -0.34]
 
 ###############################
 #parameter to map nucleotide centers to Euler translations
@@ -39,10 +40,23 @@ class oxdna_frame :
         self.base_v = -bv #minus sign for Tsukuba convention
         self.normal = n
         self.base_norv = np.cross(bv,n)
-    def int_centers(self):
-        back_bone = self.center-BB_X*self.base_v-BB_Y*self.base_norv
-        hbond = self.center+HB_X*self.base_v
-        stack = self.center+STACK_X*self.base_v
+    def int_centers(self,ty):
+        tid = 0
+        if ty == 'A' :
+            tid = 0
+        elif ty == 'G' :
+            tid = 1
+        elif ty == 'C' :
+            tid = 2
+        elif ty == "T" :
+            tid = 3
+        else :
+            print("Unknown base type " + ty)
+            quit()        
+        
+        back_bone = self.center-BB_X[tid]*self.base_v-BB_Y[tid]*self.base_norv
+        hbond = self.center+HB_X[tid]*self.base_v
+        stack = self.center+STACK_X[tid]*self.base_v
         return(back_bone,hbond,stack)
     
 ##############################  
@@ -66,11 +80,12 @@ class int_coord :
 #members are oxframe, which is an oxdna_frame object
 #and frame, which is a eframe object
 class base :
-    def __init__(self,oxc):
+    def __init__(self,oxc,ty):
         self.oxframe = oxc
+        self.type = ty
         
         #map oxdna coordinates to euler coordinates
-        p = oxc.center - OX_TZU_DISP
+        p = oxc.center + oxc.base_v*OX_TZU_DISP[0]+oxc.base_norv*OX_TZU_DISP[1]+oxc.base_norv*OX_TZU_DISP[2]
         """
         A = np.zeros((3,3), dtype=float)
         
@@ -147,7 +162,7 @@ class base_pair :
         self.frame = eframe(p,ori)
         
         #compute intra coordinates
-        rot = caym1(A2,1)
+        rot = caym1(A2,1)*180./math.pi
         tr = np.dot(self.frame.orientation.transpose(),b1.frame.pos-b2.frame.pos)
         self.intra_coord = int_coord(tr, rot)
 
@@ -166,27 +181,54 @@ class junction :
         self.frame = eframe(p,ori)
         
         #compute inter coordinates
-
-        rot = caym1(A2,1)
+        rot = caym1(A2,1)*180./math.pi
         tr = np.dot(self.frame.orientation.transpose(),bp2.frame.pos-bp1.frame.pos)
         self.inter_coord = int_coord(tr, rot)
 
 ##############################
-#read oxdna trajectory (XXXTODO: pandas?)
-#XXXTODO: read topology from topo file!
+#read oxdna trajectory
+#XXXNOTE: There is no information on base pairs in topology file + 
+#bp can potentially change runtime
+#for now I'm assuming standard order: (A)0,1,2,3,..,Nbp-1,(B)Nbp-1,Nbp-2,..,1,0 (Nbp = Nb/2 number of base pairs)  
 #XXXTODO: memory wise it's better to read AND print one snapshot at a time
[…]
