### 2022-11-22 f74380f
added stacking potential
