### 2022-11-22 f619291
added stacking potential
