### 2022-11-24 2ebcf4b
added no_center option to align
