### 2022-11-29 23cb3f2
Port two more personal interactions to the new structure
