### 2022-11-30 ef7aa63
Remove a few warnings and a call to an old CUDA-related macro
