### 2022-12-01 f574eee
Fix the way the stress tensor is computed in the interactions
