+++ contrib/rovigatti/src/Interactions/PolymerSwapInteraction.cpp
@@ -139,6 +139,15 @@ void PolymerSwapInteraction::begin_energy_computation() {
 	_chain_coms.resize(_N_chains, LR_vector(0., 0., 0.));
 
 	for(int i = 0; i < _N_chains; i++) {
+		for(auto p: CONFIG_INFO->molecules()[i]->particles) {
+			_inter_chain_stress_tensor[0] += SQR(p->vel.x);
+			_inter_chain_stress_tensor[1] += SQR(p->vel.y);
+			_inter_chain_stress_tensor[2] += SQR(p->vel.z);
+			_inter_chain_stress_tensor[3] += p->vel.x * p->vel.y;
+			_inter_chain_stress_tensor[4] += p->vel.x * p->vel.z;
+			_inter_chain_stress_tensor[5] += p->vel.y * p->vel.z;
+		}
+
 		CONFIG_INFO->molecules()[i]->update_com();
 		_chain_coms[i] = CONFIG_INFO->molecules()[i]->com;
 	}
