+++ contrib/rovigatti/src/Interactions/CUDAPolymerSwapInteraction.cu
@@ -45,6 +45,7 @@ __constant__ float MD_semiflexibility_k[1];
 
 struct __align__(16) CUDA_FS_bond {
 	c_number4 force;
+	c_number4 r;
 	int q;
 };
 
@@ -57,7 +58,7 @@ struct __align__(16) CUDA_FS_bond_list {
 					n_bonds(0) {
 	}
 	__device__
-	void add_bond(c_number4 &force, int q) {
+	void add_bond(c_number4 &force, c_number4 &r, int q) {
 		n_bonds++;
 		if(n_bonds > CUDA_MAX_SWAP_NEIGHS) {
 			printf("TOO MANY SWAP NEIGHBOURS, TRAGEDY\nHere is the list of neighbours:\n");
@@ -69,11 +70,26 @@ struct __align__(16) CUDA_FS_bond_list {
 			n_bonds--;
 		}
 		bonds[n_bonds - 1].force = force;
+		bonds[n_bonds - 1].r = r;
 		bonds[n_bonds - 1].q = q;
 	}
 };
 
-__device__ void _WCA(c_number4 &ppos, c_number4 &qpos, int int_type, c_number4 &F, CUDABox *box) {
+// in the case of pair-wise forces, which are counted twice, one should set half=true
+// while three-body forces are counted only once, so half=false should be used.
+template<bool half>
+__device__ void _update_stress_tensor(CUDAStressTensor &st, c_number4 &r, c_number4 &force) {
+	c_number factor = (half) ? 0.5f : 1.0f;
+
+	st.e[0] -= r.x * force.x * factor;
+	st.e[1] -= r.y * force.y * factor;
+	st.e[2] -= r.z * force.z * factor;
+	st.e[3] -= r.x * force.y * factor;
+	st.e[4] -= r.x * force.z * factor;
+	st.e[5] -= r.y * force.z * factor;
+}
+
+__device__ void _WCA(c_number4 &ppos, c_number4 &qpos, int int_type, c_number4 &F, CUDAStressTensor &p_st, CUDABox *box) {
 	c_number4 r = box->minimum_image(ppos, qpos);
 	c_number sqr_r = CUDA_DOT(r, r);
 
@@ -89,28 +105,24 @@ __device__ void _WCA(c_number4 &ppos, c_number4 &qpos, int int_type, c_number4 &
 		}
 		energy += 4.f * part * (part - 1.f) + 1.f - MD_alpha[0];
 		force_mod += 4.f * MD_n[0] * part * (2.f * part - 1.f) / sqr_r;
-	}
-	/*else {
-		energy += 0.5f * MD_alpha[0] * (cosf(MD_gamma[0] * sqr_r + MD_beta[0]) - 1.f);
-		force_mod += MD_alpha[0] * MD_gamma[0] * sinf(MD_gamma[0] * sqr_r + MD_beta[0]);
-	}
 
-	if(sqr_r > MD_sqr_rcut[0]) {
-		energy = force_mod = (c_number) 0.f;
-		}*/
+		c_number4 force = {-r.x * force_mod,
+				-r.y * force_mod,
+				-r.z * force_mod,
+				energy
+		};
 
-	F.x -= r.x * force_mod;
-	F.y -= r.y * force_mod;
-	F.z -= r.z * force_mod;
-	F.w += energy;
+		_update_stress_tensor<true>(p_st, r, force);
+		F += force;
+	}
 }
 
-__device__ void _sticky(c_number4 &ppos, c_number4 &qpos, int q_idx, c_number4 &F, CUDA_FS_bond_list &bond_list, CUDABox *box) {
[…]
