### 2022-12-05 d429863
Fix a bug whereby setting lastconf_file to /dev/null would segfault
