### 2022-12-05 968c5a4
Fix a bug whereby two outputs couldn't point to the same file
