### 2022-12-07 cde4d25
Make the code compilable on older gcc versions (>= 5.4)
