### 2022-12-08 bbec240
updated example input files
