### 2022-12-10 603b927
Remove unused features to compute the stress tensor in the MD backend
