### 2022-12-11 f095cd8
Add a structure to make it possible to compute the stress tensor on CUDA
Also update Str
