### 2022-12-11 fd32c23
Make Pressure and StressAutocorrelation use the new structure
