### 2022-12-12 9e41b57
Add a serialisation method to BaseObservable
The method does nothing in the base class, but gets invoked every time a
configuration is appended to the trajectory so that subclasses can
extend it
