### 2022-12-12 828879b
Strengthen up the serialisation pipeline of StressAutocorrelation
