+++ contrib/rovigatti/src/Interactions/CUDADetailedPatchySwapInteraction.cu
@@ -72,7 +72,8 @@ struct __align__(16) CUDA_FS_bond_list {
 	}
 };
 
-__device__ void _patchy_point_two_body_interaction(c_number4 &ppos, c_number4 &qpos, c_number4 &a1, c_number4 &a2, c_number4 &a3, c_number4 &b1, c_number4 &b2, c_number4 &b3, c_number4 &F, c_number4 &torque, CUDA_FS_bond_list *bonds, int q_idx, CUDABox *box) {
+__device__ void _patchy_point_two_body_interaction(c_number4 &ppos, c_number4 &qpos, c_number4 &a1, c_number4 &a2, c_number4 &a3, c_number4 &b1,
+		c_number4 &b2, c_number4 &b3, c_number4 &F, c_number4 &torque, CUDA_FS_bond_list *bonds, int q_idx, CUDAStressTensor &p_st, CUDABox *box) {
 	int ptype = get_particle_btype(ppos);
 	int qtype = get_particle_btype(qpos);
 
@@ -97,10 +98,14 @@ __device__ void _patchy_point_two_body_interaction(c_number4 &ppos, c_number4 &q
 		spherical_energy = 4.f * (SQR(lj_part) - lj_part) + 1.f - MD_spherical_attraction_strength[0];
 	}
 
-	F.x -= r.x * force_module;
-	F.y -= r.y * force_module;
-	F.z -= r.z * force_module;
-	F.w += spherical_energy - MD_spherical_E_cut[0];
+	c_number4 force = {-r.x * force_module,
+			-r.y * force_module,
+			-r.z * force_module,
+			spherical_energy - MD_spherical_E_cut[0]
+	};
+
+	_update_stress_tensor<true>(p_st, r, force);
+	F += force;
 
 	int p_N_patches = MD_N_patches[ptype];
 	int q_N_patches = MD_N_patches[qtype];
@@ -144,11 +149,12 @@ __device__ void _patchy_point_two_body_interaction(c_number4 &ppos, c_number4 &q
 
 						c_number4 p_torque = _cross(p_patch_pos, tmp_force);
 
-						torque -= p_torque;
-						F.x -= tmp_force.x;
-						F.y -= tmp_force.y;
-						F.z -= tmp_force.z;
+						F -= tmp_force;
 						F.w += energy_part;
+						torque -= p_torque;
+
+						force = -tmp_force;
+						_update_stress_tensor<true>(p_st, r, force);
 
 						CUDA_FS_bond &my_bond = bonds[p_patch].new_bond();
 
@@ -161,6 +167,7 @@ __device__ void _patchy_point_two_body_interaction(c_number4 &ppos, c_number4 &q
 							my_bond.q_torque_ref_frame = _vectors_transpose_c_number4_product(b1, b2, b3, _cross(q_patch_pos, tmp_force));
 						}
 						else {
+							my_bond.force = my_bond.p_torque = my_bond.q_torque_ref_frame = make_c_number4(0.f, 0.f, 0.f, 0.f);
 							my_bond.force.w = epsilon;
 						}
 					}
@@ -170,7 +177,8 @@ __device__ void _patchy_point_two_body_interaction(c_number4 &ppos, c_number4 &q
 	}
 }
 
-__device__ void _patchy_KF_two_body_interaction(c_number4 &ppos, c_number4 &qpos, c_number4 &a1, c_number4 &a2, c_number4 &a3, c_number4 &b1, c_number4 &b2, c_number4 &b3, c_number4 &F, c_number4 &torque, CUDA_FS_bond_list *bonds, int q_idx, CUDABox *box) {
[…]
