### 2022-12-12 ef49079
added trajectory visualization and generally cleaned up bond_analysis
