+++ analysis/src/oxDNA_analysis_tools/bond_analysis.py
@@ -4,9 +4,10 @@ from os import path
 from collections import namedtuple
 from typing import Tuple, Dict
 import numpy as np
+import matplotlib.pyplot as plt
 import oxpy
 from oxDNA_analysis_tools.UTILS.data_structures import TopInfo, TrajInfo
-from oxDNA_analysis_tools.UTILS.oat_multiprocesser import oat_multiprocesser
+from oxDNA_analysis_tools.UTILS.oat_multiprocesser import oat_multiprocesser, get_chunk_size
 from oxDNA_analysis_tools.UTILS.RyeReader import describe, get_input_parameter
 
 import time
@@ -17,7 +18,7 @@ ComputeContext = namedtuple("ComputeContext",["traj_info",
                                               "designed_pairs",
                                               "input_file"])
 
-def compute(ctx:ComputeContext, chunk_size:int, chunk_id:int):
+def compute(ctx:ComputeContext, chunk_size:int, chunk_id:int) -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
     with oxpy.Context():
         inp = oxpy.InputFile()
         inp.init_from_filename(ctx.input_file)
@@ -32,11 +33,15 @@ def compute(ctx:ComputeContext, chunk_size:int, chunk_id:int):
 
         backend = oxpy.analysis.AnalysisBackend(inp)
     
-        count_correct_bonds = 0
-        count_incorrect_bonds = 0
-        tot_bonds = 0
+        i = 0
+        count_correct_bonds = []
+        count_incorrect_bonds = []
+        tot_bonds = []
         out_array = np.zeros(ctx.top_info.nbases, dtype=int)
         while backend.read_next_configuration():
+            conf_corr_bonds = 0
+            conf_incorr_bonds = 0
+            conf_tot_bonds = 0
             pairs = backend.config_info().get_observable_by_id("my_obs").get_output_string(backend.config_info().current_step).strip().split('\n')
             for p in pairs[1:]:
                 p = p.split()
@@ -44,15 +49,19 @@ def compute(ctx:ComputeContext, chunk_size:int, chunk_id:int):
                 b = int(p[1])
                 if a in ctx.designed_pairs.keys():
                     if ctx.designed_pairs[a] == b:
-                        count_correct_bonds += 1
-                        tot_bonds += 1
+                        conf_corr_bonds += 1
+                        conf_tot_bonds += 1
                         out_array[a] += 1
                         out_array[b] += 1
                     else:
-                        count_incorrect_bonds += 1
-                        tot_bonds += 1
+                        conf_incorr_bonds += 1
+                        conf_tot_bonds += 1
 
-        return(tot_bonds, count_correct_bonds, count_incorrect_bonds, out_array)
+            count_correct_bonds.append(conf_corr_bonds)
+            count_incorrect_bonds.append(conf_incorr_bonds)
+            tot_bonds.append(conf_tot_bonds)
+
+        return(np.array(tot_bonds), np.array(count_correct_bonds), np.array(count_incorrect_bonds), out_array)
 
 
 def bond_analysis(traj_info:TrajInfo, top_info:TopInfo, pairs:Dict[int, int], inputfile:str, ncpus:int=1) -> Tuple[int, int, int, np.ndarray]:
@@ -67,33 +76,53 @@ def bond_analysis(traj_info:TrajInfo, top_info:TopInfo, pairs:Dict[int, int], in
             ncpus: (optional) number of cores to use
 
         Returns:
-            total_bonds: average number of bonds per configuraiton
-            incorrect_bonds: average number of incorrect bonds per configuration
[…]
