### 2022-12-14 4a5c31a
Add the stress tensor calculation on GPUs for DNA
