### 2022-12-15 6542715
RyeReader now handles confs with velocities removed
