+++ analysis/src/oxDNA_analysis_tools/cython_utils/get_confs.pyx
@@ -1,18 +1,26 @@
 import cython
 import numpy as np
 cimport numpy as numpy
+from sys import stderr
 from cpython.bytes cimport PyBytes_Size 
 from libc.stdio cimport fopen, fclose, fread, fseek, FILE
 from libc.string cimport strtok, strcpy
 from libc.stdlib cimport atoi, atof, atol, malloc, free
 from oxDNA_analysis_tools.UTILS.data_structures import Configuration
 
+"""
+Cython function to extract a specified number of configurations from a text trajecory file.
+
+Parameters:
+    idxs (list) : The list of starting bytes for configurations in the trajectory
+    traj_path (str) : The path to the trajectory
+    start (int) : The ID of the first configuration to read in idxs
+    nconfs (int) : How many confs to read
+    nbases (int) : How many bases per conf
+"""
 @cython.wraparound(False)
 @cython.boundscheck(False)
-def cget_confs(list idxs, str traj_path, int start, int nconfs, int nbases):
-    """
-    A string!
-    """
+def cget_confs(list idxs, str traj_path, int start, int nconfs, int nbases, bint incl_vel=1):
     # Number of configurations to read
     cdef int conf_count = len(idxs)
     if (start+nconfs >= conf_count): #this handles the last chunk which may not have nconfs confs remaining.
@@ -22,7 +30,7 @@ def cget_confs(list idxs, str traj_path, int start, int nconfs, int nbases):
     cdef int *sizes = <int *> malloc(nconfs * sizeof(int))
     cdef int *conf_starts = <int *> malloc(nconfs * sizeof(int))
     if not sizes or not conf_starts:
-        raise MemoryError("Could not allocate memory for the configuration sizes and starts")
+        raise MemoryError("ERROR: Could not allocate memory for the configuration sizes and starts", file=stderr)
     cdef int chunk_size = 0
     for i in range(nconfs):
         sizes[i] = idxs[start+i].size
@@ -36,7 +44,7 @@ def cget_confs(list idxs, str traj_path, int start, int nconfs, int nbases):
     traj_path_c[len(traj_path)] = b'\0'
     cdef FILE *traj_file = fopen(traj_path_c, "rb")
     if traj_file == NULL:
-        print("Could not open trajectory file %s" % traj_path)
+        print("ERROR: Could not open trajectory file %s" % traj_path, file=stderr)
         return
     fseek(traj_file, idxs[start].offset, 1)
 
@@ -47,7 +55,7 @@ def cget_confs(list idxs, str traj_path, int start, int nconfs, int nbases):
     # Parse the chunk into Configurations
     cdef list confs = [None]*nconfs
     for i in range(nconfs):
-        c = parse_conf(chunk, conf_starts[i], sizes[i], nbases)
+        c = parse_conf(chunk, conf_starts[i], sizes[i], nbases, incl_vel)
         confs[i] = c
 
     fclose(traj_file)
@@ -60,9 +68,9 @@ def cget_confs(list idxs, str traj_path, int start, int nconfs, int nbases):
 
 @cython.wraparound(False)
 @cython.boundscheck(False)
-cdef parse_conf(char *chunk, int start_byte, int size, int nbases):
+cdef parse_conf(char *chunk, int start_byte, int size, int nbases, bint incl_vel=1):
     cdef int THREE = 3
-    cdef long time
+    cdef numpy.int64_t time #Windows and Unix use different precision for time. Using `long` means long trajectories can't be loaded on Windows systems.
     
     #allocate some memory for our configuration
[…]
