### 2022-12-15 25148ab
Start fixing the stress tensor computation for DNA on GPUs
