+++ src/Backends/MD_CPUBackend.cpp
@@ -143,7 +143,7 @@ void MD_CPUBackend::_first_step() {
 }
 
 void MD_CPUBackend::_compute_forces() {
-	_interaction->begin_energy_computation();
+	_interaction->begin_energy_and_force_computation();
 
 	_U = (number) 0;
 	for(auto p : _particles) {
