### 2022-12-15 dad18cb
Fix two personal interactions
