+++ analysis/src/oxDNA_analysis_tools/oxDNA_PDB.py
@@ -4,107 +4,99 @@ import sys
 import os
 import numpy as np
 import copy
+import argparse
 import string
-from math import sqrt, sin
 from collections import defaultdict
-# import Bio.PDB as bio
+from math import sqrt, sin
 
 from oxDNA_analysis_tools.UTILS.pdb import Atom, Nucleotide, AminoAcid, FROM_OXDNA_TO_ANGSTROM
-
+from oxDNA_analysis_tools.UTILS.RyeReader import get_confs, describe, strand_describe, inbox
 import oxDNA_analysis_tools.UTILS.protein_to_pdb as pro
-import oxDNA_analysis_tools.UTILS.base as base
 import oxDNA_analysis_tools.UTILS.utils as utils
 
-from oxDNA_analysis_tools.UTILS.readers import LorenzoReader2
-
 DD12_PDB_PATH = "./UTILS/dd12_na.pdb"
-protein_pdb_files = []
-def print_usage():
-        print("USAGE:", file=sys.stderr)
-        print("\t%s topology configuration direction pdbfiles(optional, needs flag -p) " % sys.argv[0], file =sys.stderr)
-        print("\t[-H\--hydrogens=True] [-u\--uniform-residue-names] [-p\--contains-protein] [-r\--rmsf-file]", file=sys.stderr)
-        exit(1)
 
-def parse_options():
-    shortArgs = 'H:uospr'
-    longArgs = ['hydrogens=','uniform-residue-names', 'one-file-per-strand', 'contains-protein', 'rmsf-file']
-    
-    opts = {
-        "configuration" : "",
-        "topology" : "",
-        "oxDNA_direction" : True,
-        "print_hydrogens" : True,
-        "uniform_residue_names" : False,
-        "one_file_per_strand" : False,
-        "contains_protein" : False,
-        "rmsf_bfactor": "",
-    }
-    
-    try:
-        import getopt
-        args, positional_args = getopt.gnu_getopt(sys.argv[1:], shortArgs, longArgs)
-        for k in args:
-            if k[0] == '-H' or k[0] == '--hydrogens':
-                k_arg = k[1].lower()
-                if k_arg.lower() == "true":
-                    opts["print_hydrogens"] = True
-                elif k_arg == "false":
-                    print >> sys.stderr, "## Hydrogen atoms will *not* be printed"
-                    opts["print_hydrogens"] = False
-                else:
-                    print >> sys.stderr, "The argument of '%s' should be either 'true' or 'false' (got '%s' instead)" % (k[0], k[1])
-                    exit(1)
-            elif k[0] == '-u' or k[0] == '--uniform-residue-names':
-                opts["uniform_residue_names"] = True
-            elif k[0] == '-o' or k[0] == '--one-file-per-strand':
-                opts["one_file_per_strand"] = True
-            elif k[0] == '-s' or k[0] == '--same-pdb-all-protein-strands':
-                opts["same_pdb_all_protein_strands"] = True
-            elif k[0] == '-p' or k[0] == '--contains-protein':
-                opts["contains_protein"] = True
-            elif k[0] == '-r' or k[0] == '--rmsf-file':
-                opts["rmsf_bfactor"] = k[1]
-            
-        opts['topology'] = positional_args[0]
-        opts['configuration'] = positional_args[1]
-        direction = positional_args[2]
-        if opts["contains_protein"]:
-            protein_pdb_files = positional_args[3:]
-        else:
-            protein_pdb_files=[]
-        
-        if direction == "35":
-            opts["oxDNA_direction"] = True
-        elif direction == "53":
-            opts["oxDNA_direction"] = False
-        else:
[…]
