### 2022-12-19 ead6073
Created file_info.py to get trajectory info
