### 2022-12-22 7c115cc
don't write momenta in RyeConfs
