### 2023-01-07 b332542
Improve CUDA performance by removing unused members from a struct
