### 2023-01-12 9e93ab0
Improve the performance of the DNA and RNA interactions
This changes adds some const and __restrict__ keywords to kernel and
device function arguments. It also improves the overall performances
when use_edge = true and edge_n_forces = 1 (which is the standard for
DNA and RNA simulations).
