### 2023-01-13 13dea2c
Fix the CUDA GPU interaction and improve its performance by 10-15%
