+++ src/CUDA/Interactions/CUDADNAInteraction.h
@@ -38,6 +38,7 @@ public:
 	float _debye_huckel_B; // prefactor of the quadratic cut-off
 	float _minus_kappa;
 	// End copy from DNA2Interaction.h
+	int *_d_is_strand_end = nullptr;
 
 	void get_settings(input_file &inp) override;
 	void cuda_init(int N) override;
@@ -45,12 +46,14 @@ public:
 		return this->get_rcut();
 	}
 
-	void _on_T_update() override;
-
 	void compute_forces(CUDABaseList*lists, c_number4 *d_poss, GPU_quat *d_qorientations, c_number4 *d_forces, c_number4 *d_torques, LR_bonds *d_bonds, CUDABox*d_box);
 	void _hb_op_precalc(c_number4 *poss, GPU_quat *orientations, int *op_pairs1, int *op_pairs2, float *hb_energies, int n_threads, bool *region_is_nearhb, CUDA_kernel_cfg hb_kernel_cfg, CUDABox*d_box);
 	void _near_hb_op_precalc(c_number4 *poss, GPU_quat *orientations, int *op_pairs1, int *op_pairs2, bool *nearly_bonded_array, int n_threads, bool *region_is_nearhb, CUDA_kernel_cfg hb_kernel_cfg, CUDABox*d_box);
 	void _dist_op_precalc(c_number4 *poss, GPU_quat *orientations, int *op_pairs1, int *op_pairs2, c_number *op_dists, int n_threads, CUDA_kernel_cfg dist_kernel_cfg, CUDABox*d_box);
+
+protected:
+	void _on_T_update() override;
+	void _init_strand_ends(LR_bonds *d_bonds);
 };
 
 #endif /* CUDADNAINTERACTION_H_ */
