### 2023-01-14 de4e47e
Remove a warning and add some minor performance improvements
