### 2023-01-16 #4693 @degiacom
I apologise for having taken this long. As @ppxasjsm mentioned, I had some problems with installation, that turned out to be my fault. I told cmake to use this compiler via the flags `CMAKE_C_COMPILER` and `CMAKE_CXX_COMPILER`, but oddly things did not work. Compilation with g++ version 5.4 worked smoothly by setting environment variables `CC` and `CXX` though. No problem in installing Oxpy too.

The examples folder is very useful. I have only two minor comments about it.

- it appears that previously a folder called `UTILS` existed. This seems to have been moved to `legacy/UTILS`, a folder that also feature scripts using Python 2 syntax that will break on computer using updated Python versions. A `utils` (lowercase) folder is now available, but it does not contain most of the files in its uppercase counterpart.  As a result, some of the examples, either code or readme files, may have to be updated (I suggest grepping to find the list of all the occurrences). Examples:
    - The README file in the example folder `examples/CIRC_ASSEMBLY` indicates that `../../UTILS/generate-sa.py 40. sequences` should be run. The Python script can be found, and correctly executed, in `utils`.
    - the script `examples/MAJOR_MINOR_GROOVING/traj2chimera_grooving.sh` breaks, as it attempts calling .`./../UTILS/traj2chimera.py`. The latter script is not available in the `utils` folder, but appears to be available in `legacy/UTILS`. However, this calls methods from `legacy/UTILS/base.py`, that appears to be using Python 2 syntax (`exec` instead of `exec( )`).

- in the README file of `examples/CIRC_ASSEMBLY` I noticed that the user is prompted to run `../../bin/oxDNA` input. If the user has installed oxDNA as per the installation instructions though, the executable will be in `../../build/bin/oxDNA`. A grep reveals that this happens in few more examples (i.e., the executables within `examples/FFS_example` and `examples/RELAX_INITIAL_CONFIGURATION`).

Besides those two minor points, I think this is an excellent piece of software. I congratulate the developers for their work, and again apologise for having taken so long.
