### 2023-01-16 dfcd63a
Start fixing some of the issues reported in the joss review
The review is
[here](https://github.com/openjournals/joss-reviews/issues/4693#issuecomment-1384241761)
