### 2023-01-16 1874b66
Remove some outdated examples and update others
