+++ examples/PERSISTENCE_LENGTH/readme.tex
@@ -16,7 +16,7 @@ Then you need to run
  oxDNA input_persistence
 \end{verbatim}
 
-Note that for caclulation of persistence length, one needs a large number of decorrelated states. The program will produce a trajectory.dat file. To analyze the data, use the python script 
+Note that for calculation of persistence length, one needs a large number of decorrelated states. The program will produce a trajectory.dat file. To analyze the data, use the python script 
 dspl.py:
 \begin{verbatim}
  dspl.py trajectory.dat init.top 10 50
