### 2023-01-17 cf03564
Fix typo in file_info.py
