### 2023-01-17 bf40c14
untested rewrite of oxpy_remd's reshuffle.py
