### 2023-01-18 0112a3e
Fix two bugs in the CUDARNA interaction and slightly improve performance
