+++ src/CUDA/Interactions/CUDARNAInteraction.cu
@@ -223,7 +223,9 @@ CUDARNAInteraction::CUDARNAInteraction() {
 }
 
 CUDARNAInteraction::~CUDARNAInteraction() {
-
+	if(_d_is_strand_end != nullptr) {
+		CUDA_SAFE_CALL(cudaFree(_d_is_strand_end));
+	}
 }
 
 void CUDARNAInteraction::get_settings(input_file &inp) {
@@ -235,7 +237,7 @@ void CUDARNAInteraction::get_settings(input_file &inp) {
 			_use_debye_huckel = true;
 			// copy-pasted from the DNA2Interaction constructor
 			_debye_huckel_half_charged_ends = true;
-			this->_grooving = true;
+			_grooving = true;
 			// end copy from DNA2Interaction
 
 			// copied from DNA2Interaction::get_settings() (CPU), the least bad way of doing things
@@ -253,15 +255,15 @@ void CUDARNAInteraction::get_settings(input_file &inp) {
 
 			int mismatches = 0;
 			if(getInputBoolAsInt(&inp, "mismatch_repulsion", &mismatches, 0) == KEY_FOUND) {
-				this->_mismatch_repulsion = (bool) mismatches;
+				_mismatch_repulsion = (bool) mismatches;
 			}
-			if(this->_mismatch_repulsion) {
+			if(_mismatch_repulsion) {
 				float temp;
 				if(getInputFloat(&inp, "mismatch_repulsion_strength", &temp, 0) == KEY_FOUND) {
-					this->_RNA_HYDR_MIS = temp;
+					_RNA_HYDR_MIS = temp;
 				}
 				else {
-					this->_RNA_HYDR_MIS = 1;
+					_RNA_HYDR_MIS = 1;
 				}
 
 			}
@@ -276,23 +278,23 @@ void CUDARNAInteraction::cuda_init(int N) {
 	CUDABaseInteraction::cuda_init(N);
 	RNAInteraction::init();
 
-	float f_copy = 1.0; //this->_hb_multiplier;
+	float f_copy = 1.0; //_hb_multiplier;
 	CUDA_SAFE_CALL(cudaMemcpyToSymbol(MD_hb_multi, &f_copy, sizeof(float)));
 
 	CUDA_SAFE_CALL(cudaMemcpyToSymbol(MD_N, &N, sizeof(int)));
 
 	//mismatch repulsion modification
-	if(this->_mismatch_repulsion) {
-		float tempmis = -1.0 * this->_RNA_HYDR_MIS / this->model->RNA_HYDR_EPS;
-		this->F1_EPS[RNA_HYDR_F1][0][0] *= tempmis;
-		this->F1_SHIFT[RNA_HYDR_F1][0][0] *= tempmis;
+	if(_mismatch_repulsion) {
+		float tempmis = -1.0 * _RNA_HYDR_MIS / model->RNA_HYDR_EPS;
+		F1_EPS[RNA_HYDR_F1][0][0] *= tempmis;
+		F1_SHIFT[RNA_HYDR_F1][0][0] *= tempmis;
 	}
 
 	c_number tmp[50];
 	for(int i = 0; i < 2; i++) {
 		for(int j = 0; j < 5; j++) {
 			for(int k = 0; k < 5; k++) {
-				tmp[i * 25 + j * 5 + k] = this->F1_EPS[i][j][k];
[…]
