+++ src/CUDA/Interactions/CUDADNAInteraction.cu
@@ -130,7 +130,7 @@ void CUDADNAInteraction::cuda_init(int N) {
 
 		c_number debyecut;
 		if(this->_grooving) {
-			debyecut = 2.0f * sqrt((POS_MM_BACK1) * (POS_MM_BACK1) + (POS_MM_BACK2) * (POS_MM_BACK2)) + _debye_huckel_RC;
+			debyecut = 2.0f * sqrt(SQR(POS_MM_BACK1) + SQR(POS_MM_BACK2)) + _debye_huckel_RC;
 		}
 		else {
 			debyecut = 2.0f * sqrt(SQR(POS_BACK)) + _debye_huckel_RC;
@@ -157,7 +157,7 @@ void CUDADNAInteraction::_on_T_update() {
 
 void CUDADNAInteraction::_init_strand_ends(LR_bonds *d_bonds) {
 	CUDA_SAFE_CALL(GpuUtils::LR_cudaMalloc<int>(&_d_is_strand_end, sizeof(int) * _N));
-	init_strand_ends<<<_launch_cfg.blocks, _launch_cfg.threads_per_block>>>(_d_is_strand_end, d_bonds);
+	init_DNA_strand_ends<<<_launch_cfg.blocks, _launch_cfg.threads_per_block>>>(_d_is_strand_end, d_bonds, _N);
 }
 
 void CUDADNAInteraction::compute_forces(CUDABaseList *lists, c_number4 *d_poss, GPU_quat *d_orientations, c_number4 *d_forces, c_number4 *d_torques, LR_bonds *d_bonds, CUDABox*d_box) {
