### 2023-01-19 ffbae41
read two contributes H,S from matrix.dat
