### 2023-01-19 2765c53
Make the code compilable on older gcc versions (>= 5.4)
