### 2023-01-19 6c51446
Make the code compilable on older gcc versions (>= 5.4)
