+++ src/Forces/Metadynamics/LTCOMAngleTrap.cpp
@@ -58,13 +58,13 @@ std::tuple<std::vector<int>, std::string> LTCOMAngleTrap::init(input_file &inp) 
 
 	std::string description = Utils::sformat("LTCOMAngleTrap force with mode = %d", _mode);
 	if(_mode == 1) {
-		return {_p1a, description};
+		return std::make_tuple(_p1a, description);
 	}
 	else if(_mode == 2) {
-		return {_p2a, description};
+		return std::make_tuple(_p2a, description);
 	}
 	else {
-		return {_p3a, description};
+		return std::make_tuple(_p3a, description);
 	}
 }
 
