### 2023-01-19 b7f090e
Make the code compilable on older gcc versions (>= 5.4)
