+++ analysis/src/oxDNA_analysis_tools/persistence_length.py
@@ -1,0 +1,219 @@
+import sys
+from os import path
+from typing import Dict, Tuple
+from collections import namedtuple
+import argparse
+import numpy as np
+import matplotlib.pyplot as plt
+import oxpy
+from oxDNA_analysis_tools.UTILS.RyeReader import describe
+from oxDNA_analysis_tools.UTILS.oat_multiprocesser import oat_multiprocesser
+from oxDNA_analysis_tools.UTILS.data_structures import TrajInfo
+
+import time
+start_time = time.time()
+
+ComputeContext = namedtuple("ComputeContext",["traj_info",
+                                              "input_file",
+                                              "n1",
+                                              "n2"])
+
+def get_r(conf, nucid:int, pair_dict:Dict) -> np.ndarray:
+    """
+        Returns a normalized vector pointing from base midpoint of the `nucid`-th base pair to the midpoint of the next base pair.
+        
+        Parameters:
+            conf (oxpy.config_info) : The current configuration
+            nucid (int) : ID of the nucleotide in the first strand to compute the vector from
+
+        Returns:
+            (np.ndarray) : The vector pointing from the midpoint of nucid's base pair to the +1 base pair.
+    """
+    box = np.array(conf.box.box_sides)
+    pair = pair_dict[nucid]
+    next_pair = pair_dict[nucid+1]
+
+    firstA = conf.particles()[nucid].base_site()
+    firstB = conf.particles()[pair].base_site()
+    secondA = conf.particles()[nucid+1].base_site()
+    secondB = conf.particles()[next_pair].base_site()
+
+    first_midpos = (firstA + firstB) / 2
+    second_midpos = (secondA + secondB) / 2 
+
+    r = second_midpos - first_midpos 		
+    r -= box * np.rint(r / box)
+    return r
+
+def compute(ctx:ComputeContext, chunk_size:int, chunk_id:int):
+    l0 = 0
+    correlations = np.zeros(ctx.n2 - ctx.n1)
+    correlations_counter = np.zeros_like(correlations)
+    # Read in a conf and identify h-bonds
+    with oxpy.Context():
+        inp = oxpy.InputFile()
+        inp.init_from_filename(ctx.input_file)
+        inp["list_type"] = "cells"
+        inp["trajectory_file"] = ctx.traj_info.path
+        inp["analysis_bytes_to_skip"] = str(ctx.traj_info.idxs[chunk_id*chunk_size].offset)
+        inp["confs_to_analyse"] = str(chunk_size)
+        inp["analysis_data_output_1"] = '{ \n name = stdout \n print_every = 1e10 \n col_1 = { \n id = my_obs \n type = hb_list \n } \n }'
+
+        if (not inp["use_average_seq"] or inp.get_bool("use_average_seq")) and "RNA" in inp["interaction_type"]:
+            print("WARNING: Sequence dependence not set for RNA model, wobble base pairs will be ignored", file=sys.stderr)
+
+        backend = oxpy.analysis.AnalysisBackend(inp)
+        if backend.config_info().particles()[ctx.n1].strand_id != backend.config_info().particles()[ctx.n2].strand_id:
+            print("ERROR nucid_1 and nucid_2 must be on the same strand", file=sys.stderr)
+            sys.exit()
+
+        while backend.read_next_configuration():
+            pairs = backend.config_info().get_observable_by_id("my_obs").get_output_string(backend.config_info().current_step).strip().split('\n')[1:]
+
+            # Extract paired nucleotides into a dict
+            pair_dict = {}
+            for p in pairs:
+                p1 = int(p.split()[0])
+                p2 = int(p.split()[1])
+                if p1 < ctx.n1 or p1 > ctx.n2:
+                    continue
[…]
