### 2023-01-19 4b25590
Fix the oxDNA exe path in some examples
This is in response to one of the issues raised during the [JOSS
review](https://github.com/openjournals/joss-reviews/issues/4693#issuecomment-1384241761)
