+++ examples/RELAX_INITIAL_CONFIGURATION/run.sh
@@ -2,7 +2,7 @@
 
 # run the relaxation procedure
 
-../../bin/oxDNA input_weak_constant
+../../build/bin/oxDNA input_weak_constant
 mv last_conf.dat relaxed_weak.conf
-../../bin/oxDNA input_strong_constant
+../../build/bin/oxDNA input_strong_constant
 mv last_conf.dat relaxed.conf
