### 2023-01-20 #4693 @ppxasjsm
Whoops sorry my bad, small thing is still missing. Can you please generate a tagged release of the software (a new one if changed), and archive it (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or other). Please then let me know the version number and the archive DOI in this issue here.
