### 2023-01-20 #4693 @lorenzo-rovigatti
Sure! The release is [here](https://github.com/lorenzo-rovigatti/oxDNA/releases/tag/v3.5.0), while the DOI is: [![DOI](https://zenodo.org/badge/209079824.svg)](https://zenodo.org/badge/latestdoi/209079824)
