### 2023-01-21 #4693 @Kevin-Mattheus-Moerman
@lorenzo-rovigatti I am the AEiC for this track and here to help process the last required steps for publication in JOSS.

On the paper:
- [ ] Please fix those potentially missing/invalid DOIs ☝️
- [ ] Please spell out all country names, so do not use acronyms like USA and UK.
- [ ] The words `modernized` and `idealized` are American English, while `behaviour` and `analyse` are Brittish English. I suggest you choose one and adjust spelling throughout accordingly (I simply pasted your raw paper.md text into a text editor with a spelling checker to find these, so you could do the same to find/fix them if you like).

On the archive:
- [ ] Please manually edit ZENODO archive title to match the paper title
- [ ] Please also manually edit the ZENODO archive author list to match the paper authors and order (you may need to remove any contributors that were added on the ZENODO list)
- [ ] Please manually edit the ZENODO listed license (currently it says `Other (Open)`) to match your actual software's open source license.
